Finish audit test robustness follow-up
