+++ src/roomeq/optimize/tests.rs
@@ -310,1805 +310,8 @@ fn observer_none() -> SharedPipelineObserver {
     Arc::new(Mutex::new(None))
 }
 
-#[test]
-fn select_topology_route_stereo_2_0() {
-    let speakers = stereo_speakers();
-    let system = SystemConfig {
-        model: SystemModel::Stereo,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::Stereo2_0);
-}
-
-#[test]
-fn select_topology_route_stereo_2_1() {
-    let mut speakers = stereo_speakers();
-    speakers.insert(
-        "sub".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::Stereo,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-            ("Sub".to_string(), "sub".to_string()),
-        ]),
-        subwoofers: Some(SubwooferSystemConfig {
-            config: SubwooferStrategy::Single,
-            crossover: None,
-            mapping: [("sub".to_string(), "Left".to_string())].into(),
-        }),
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::Stereo2_1);
-}
-
-#[test]
-fn select_topology_route_home_cinema_with_sub() {
-    let mut speakers = stereo_speakers();
-    speakers.insert(
-        "center".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "lfe".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::HomeCinema,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-            ("Center".to_string(), "center".to_string()),
-            ("LFE".to_string(), "lfe".to_string()),
-        ]),
-        subwoofers: Some(SubwooferSystemConfig {
-            config: SubwooferStrategy::Single,
-            crossover: None,
-            mapping: [("lfe".to_string(), "Center".to_string())].into(),
-        }),
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::HomeCinema);
-}
-
-#[test]
-fn select_topology_route_home_cinema_without_sub() {
-    let mut speakers = stereo_speakers();
-    speakers.insert(
-        "center".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::HomeCinema,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-            ("Center".to_string(), "center".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::HomeCinema);
-}
-
-#[test]
-fn select_topology_route_custom_is_generic() {
-    let speakers = stereo_speakers();
-    let system = SystemConfig {
-        model: SystemModel::Custom,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::Generic);
-}
-
-#[test]
-fn select_topology_route_no_system_is_generic() {
-    let speakers = stereo_speakers();
-    let config = base_room_config(speakers, None);
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::Generic);
-}
-
-#[test]
-fn select_topology_route_speaker_group_falls_back_to_generic() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left_group".to_string(),
-        SpeakerConfig::Group(SpeakerGroup {
-            name: "left_group".to_string(),
-            speaker_name: None,
-            measurements: vec![MeasurementSource::InMemory(flat_curve())],
-            crossover: None,
-        }),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::Stereo,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left_group".to_string()),
-            ("Right".to_string(), "right".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system));
-    let route = select_topology_route(&config, &observer_none()).unwrap();
-    assert_eq!(route, TopologyRoute::Generic);
-}
-
-#[test]
-fn validate_room_optimization_empty_speakers_fails() {
-    let config = RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: None,
-        speakers: HashMap::new(),
-        crossovers: None,
-        target_curve: None,
-        optimizer: OptimizerConfig::default(),
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    };
-    let observer = observer_none();
-    let result = validate_room_optimization(&config, &observer);
-    assert!(result.is_err(), "empty speakers should fail validation");
-}
-
-fn stereo_2_0_config() -> RoomConfig {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-
-    RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: Some(SystemConfig {
-            model: SystemModel::Stereo,
-            speakers: HashMap::from([
-                ("Left".to_string(), "left".to_string()),
-                ("Right".to_string(), "right".to_string()),
-            ]),
-            subwoofers: None,
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: None,
-        target_curve: None,
-        optimizer: OptimizerConfig {
-            processing_mode: ProcessingMode::LowLatency,
-            num_filters: 1,
-            max_iter: 20,
-            population: 6,
-            min_freq: 20.0,
-            max_freq: 500.0,
-            psychoacoustic: false,
-            refine: false,
-            ..Default::default()
-        },
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    }
-}
-
-#[test]
-fn execute_topology_workflow_stereo_2_0_returns_result() {
-    let config = stereo_2_0_config();
-    let sys = config.system.as_ref().unwrap();
-    let observer = observer_none();
-    let route = TopologyRoute::Stereo2_0;
-    let result = execute_topology_workflow(&config, sys, 48000.0, None, &observer, route).unwrap();
-    assert!(!result.channels.is_empty(), "result should have channels");
-}
-
-#[test]
-fn execute_topology_workflow_home_cinema_without_sub() {
-    let mut speakers = stereo_speakers();
-    speakers.insert(
-        "center".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: Some(SystemConfig {
-            model: SystemModel::HomeCinema,
-            speakers: HashMap::from([
-                ("Left".to_string(), "left".to_string()),
-                ("Right".to_string(), "right".to_string()),
-                ("Center".to_string(), "center".to_string()),
-            ]),
-            subwoofers: None,
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: None,
-        target_curve: None,
-        optimizer: OptimizerConfig {
-            processing_mode: ProcessingMode::LowLatency,
-            num_filters: 1,
-            max_iter: 20,
-            population: 6,
-            min_freq: 20.0,
-            max_freq: 500.0,
-            psychoacoustic: false,
-            refine: false,
-            ..Default::default()
-        },
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    };
-    let sys = config.system.as_ref().unwrap();
-    let observer = observer_none();
-    let route = TopologyRoute::HomeCinema;
-    let result = execute_topology_workflow(&config, sys, 48000.0, None, &observer, route).unwrap();
-    assert!(!result.channels.is_empty(), "result should have channels");
-}
-
-#[test]
-fn execute_generic_channels_single_speaker() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer = observer_none();
-    let (generic, total_speakers) =
-        execute_generic_channels(&config, 48000.0, None, None, &observer).unwrap();
-    assert_eq!(total_speakers, 1);
-    assert!(
-        generic.channel_results.contains_key("left"),
-        "generic results should contain 'left'"
-    );
-}
-
-#[test]
-fn assemble_workflow_result_persists_channels() {
-    let speakers = stereo_speakers();
-    let system = SystemConfig {
-        model: SystemModel::Stereo,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = base_room_config(speakers, Some(system.clone()));
-    let result = single_channel_room_result("left");
-    let sys = config.system.as_ref().unwrap();
-
-    let assembled = assemble_workflow_result(
-        result,
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer_none(),
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        assembled.is_ok(),
-        "workflow assembly should succeed: {:?}",
-        assembled.err()
-    );
-    let assembled = assembled.unwrap();
-    assert!(
-        assembled.channel_results.contains_key("left"),
-        "assembled result should preserve channel results"
-    );
-}
-
-#[test]
-fn assemble_generic_result_empty_channels_fails() {
-    let config = base_room_config(HashMap::new(), None);
-    let generic = GenericChannelCollection {
-        channel_chains: HashMap::new(),
-        channel_results: HashMap::new(),
-        pre_scores: Vec::new(),
-        post_scores: Vec::new(),
-        curves: HashMap::new(),
-        channel_means: HashMap::new(),
-        channel_arrivals: HashMap::new(),
-    };
-
-    let assembled = assemble_generic_result(
-        generic,
-        0,
-        &config,
-        48000.0,
-        None,
-        &observer_none(),
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        assembled.is_err() || assembled.as_ref().unwrap().channel_results.is_empty(),
-        "generic assembly with no channels should error or produce no channels"
-    );
-}
-
-#[test]
-fn sanity_check_result_empty_channels_errors() {
-    let result = RoomOptimizationResult {
-        channels: HashMap::new(),
-        channel_results: HashMap::new(),
-        combined_pre_score: 0.0,
-        combined_post_score: 0.0,
-        metadata: empty_metadata(),
-    };
-    assert!(
-        sanity_check_result(&result).is_err(),
-        "sanity check should fail when no channels are produced"
-    );
-}
-
-// =============================================================================
-// Branch coverage for decomposed optimize_room pipeline
-// =============================================================================
-
-fn tiny_optimizer() -> OptimizerConfig {
-    OptimizerConfig {
-        processing_mode: ProcessingMode::LowLatency,
-        num_filters: 1,
-        max_iter: 20,
-        population: 6,
-        seed: Some(1),
-        min_freq: 20.0,
-        max_freq: 500.0,
-        psychoacoustic: false,
-        refine: false,
-        ..Default::default()
-    }
-}
-
-fn room_config_with_optimizer(
-    speakers: HashMap<String, SpeakerConfig>,
-    system: Option<SystemConfig>,
-    optimizer: OptimizerConfig,
-) -> RoomConfig {
-    RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system,
-        speakers,
-        crossovers: None,
-        target_curve: None,
-        optimizer,
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    }
-}
-
-// ---- prepare_room_optimization ------------------------------------------------
-
-#[test]
-fn prepare_room_optimization_without_system_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-
-    let (observer, prepared) = prepare_room_optimization(&config, None).unwrap();
-    assert!(prepared.system.is_none());
-    assert!(observer.lock().unwrap().is_none());
-}
-
-#[test]
-fn prepare_room_optimization_with_system_and_bass_management_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::HomeCinema,
-        speakers: HashMap::from([("Left".to_string(), "left".to_string())]),
-        subwoofers: None,
-        bass_management: Some(BassManagementConfig {
-            enabled: true,
-            ..BassManagementConfig::default()
-        }),
-        ..Default::default()
-    };
-    let config = room_config_with_optimizer(speakers, Some(system), tiny_optimizer());
-
-    let (_observer, prepared) = prepare_room_optimization(&config, None).unwrap();
-    assert!(
-        prepared
-            .system
-            .as_ref()
-            .unwrap()
-            .bass_management
-            .as_ref()
-            .unwrap()
-            .enabled
-    );
-}
-
-// ---- validate_room_optimization error paths ----------------------------------
-
-#[test]
-fn validate_room_optimization_mismatched_multi_measurement_weights_fails() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemoryMultiple(vec![
-            flat_curve(),
-            flat_curve(),
-        ])),
-    );
-    let mut optimizer = tiny_optimizer();
-    optimizer.multi_measurement = Some(MultiMeasurementConfig {
-        strategy: MultiMeasurementStrategy::WeightedSum,
-        weights: Some(vec![1.0]), // one weight for two measurements
-        ..MultiMeasurementConfig::default()
-    });
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-
-    let result = validate_room_optimization(&config, &observer_none());
-    assert!(
-        result.is_err(),
-        "mismatched multi-measurement weights should fail validation: {:?}",
-        result
-    );
-}
-
-#[test]
-fn validate_room_optimization_missing_measurements_fails() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Group(SpeakerGroup {
-            name: "left".to_string(),
-            speaker_name: None,
-            measurements: vec![],
-            crossover: None,
-        }),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-
-    let result = validate_room_optimization(&config, &observer_none());
-    assert!(
-        result.is_err(),
-        "speaker group with no measurements should fail validation: {:?}",
-        result
-    );
-}
-
-// ---- select_topology_route / invalid topology --------------------------------
-
-#[test]
-fn execute_topology_workflow_generic_route_errors() {
-    let config = stereo_2_0_config();
-    let sys = config.system.as_ref().unwrap();
-    let result = execute_topology_workflow(
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer_none(),
-        TopologyRoute::Generic,
-    );
-    assert!(
-        result.is_err(),
-        "generic route should not be handled by execute_topology_workflow"
-    );
-}
-
-// ---- execute_topology_workflow happy paths and early stop --------------------
-
-fn stereo_2_1_config() -> RoomConfig {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "sub".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-
-    let mut crossovers = HashMap::new();
-    crossovers.insert(
-        "xo".to_string(),
-        CrossoverConfig {
-            crossover_type: "LR24".to_string(),
-            frequency: Some(80.0),
-            frequencies: None,
-            frequency_range: None,
-        },
-    );
-
-    RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: Some(SystemConfig {
-            model: SystemModel::Stereo,
-            speakers: HashMap::from([
-                ("L".to_string(), "left".to_string()),
-                ("R".to_string(), "right".to_string()),
-                ("Sub".to_string(), "sub".to_string()),
-            ]),
-            subwoofers: Some(SubwooferSystemConfig {
-                config: SubwooferStrategy::Single,
-                crossover: Some("xo".to_string()),
-                mapping: [("sub".to_string(), "L".to_string())].into(),
-            }),
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: Some(crossovers),
-        target_curve: None,
-        optimizer: tiny_optimizer(),
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    }
-}
-
-#[test]
-fn execute_topology_workflow_stereo_2_1_returns_result() {
-    let config = stereo_2_1_config();
-    let sys = config.system.as_ref().unwrap();
-    let result = execute_topology_workflow(
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer_none(),
-        TopologyRoute::Stereo2_1,
-    )
-    .unwrap();
-    assert!(
-        !result.channels.is_empty(),
-        "stereo 2.1 should produce channels"
-    );
-}
-
-#[test]
-fn execute_topology_workflow_stops_on_observer_request() {
-    let config = stereo_2_0_config();
-    let sys = config.system.as_ref().unwrap();
-    let observer: SharedPipelineObserver = Arc::new(Mutex::new(Some(Box::new(
-        |_event: &PipelineEvent| -> PipelineControl { PipelineControl::Stop },
-    ))));
-    let result = execute_topology_workflow(
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer,
-        TopologyRoute::Stereo2_0,
-    );
-    assert!(
-        result.is_err(),
-        "workflow should halt when observer requests stop"
-    );
-}
-
-// ---- execute_generic_channels happy path and early stop ----------------------
-
-#[test]
-fn execute_generic_channels_stops_on_observer_request() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let observer: SharedPipelineObserver = Arc::new(Mutex::new(Some(Box::new(
-        |_event: &PipelineEvent| -> PipelineControl { PipelineControl::Stop },
-    ))));
-
-    let result = execute_generic_channels(&config, 48000.0, None, None, &observer);
-    assert!(
-        result.is_err(),
-        "generic channel optimization should halt when observer requests stop"
-    );
-}
-
-// ---- assemble_workflow_result / assemble_generic_result ----------------------
-
-#[test]
-fn assemble_workflow_result_empty_channels_succeeds() {
-    let speakers = stereo_speakers();
-    let system = SystemConfig {
-        model: SystemModel::Stereo,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = room_config_with_optimizer(speakers, Some(system), tiny_optimizer());
-    let sys = config.system.as_ref().unwrap();
-    let result = RoomOptimizationResult {
-        channels: HashMap::new(),
-        channel_results: HashMap::new(),
-        combined_pre_score: 0.0,
-        combined_post_score: 0.0,
-        metadata: empty_metadata(),
-    };
-
-    let assembled = assemble_workflow_result(
-        result,
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer_none(),
-        &crate::MemoryArtifactStore::new(),
-    )
-    .unwrap();
-    assert!(assembled.channel_results.is_empty());
-}
-
-#[test]
-fn assemble_generic_result_non_empty_success() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-
-    let curve = flat_curve();
-    let mut channel_chains = HashMap::new();
-    channel_chains.insert(
-        "left".to_string(),
-        ChannelDspChain {
-            channel: "left".to_string(),
-            plugins: Vec::new(),
-            drivers: None,
-            initial_curve: None,
-            final_curve: None,
-            eq_response: None,
-            target_curve: None,
-            pre_ir: None,
-            post_ir: None,
-            fir_temporal_masking: None,
-            direct_early_late_correction: None,
-        },
-    );
-    let mut channel_results = HashMap::new();
-    channel_results.insert(
-        "left".to_string(),
-        ChannelOptimizationResult {
-            name: "left".to_string(),
-            pre_score: 0.5,
-            post_score: 0.9,
-            initial_curve: curve.clone(),
-            final_curve: curve,
-            biquads: Vec::new(),
-            fir_coeffs: None,
-        },
-    );
-    let mut curves = HashMap::new();
-    curves.insert("left".to_string(), flat_curve());
-    let mut channel_means = HashMap::new();
-    channel_means.insert("left".to_string(), 80.0);
-
-    let generic = GenericChannelCollection {
-        channel_chains,
-        channel_results,
-        pre_scores: vec![0.5],
-        post_scores: vec![0.9],
-        curves,
-        channel_means,
-        channel_arrivals: HashMap::new(),
-    };
-
-    let assembled = assemble_generic_result(
-        generic,
-        1,
-        &config,
-        48000.0,
-        None,
-        &observer_none(),
-        &crate::MemoryArtifactStore::new(),
-    )
-    .unwrap();
-    assert!(
-        assembled.channel_results.contains_key("left"),
-        "assembled generic result should preserve channel results"
-    );
-}
-
-// ---- sanity_check_result -----------------------------------------------------
-
-#[test]
-fn sanity_check_result_non_empty_ok() {
-    let result = single_channel_room_result("left");
-    assert!(
-        sanity_check_result(&result).is_ok(),
-        "sanity check should pass for non-empty results"
-    );
-}
-
-// ---- optimize_room_impl callback / observer / probe-arrivals -----------------
-
-#[test]
-fn optimize_room_impl_without_probe_arrivals_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "optimize_room_impl without probe arrivals should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_impl_with_probe_arrivals_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let mut probe = HashMap::new();
-    probe.insert("left".to_string(), 5.0);
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        Some(&probe),
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "optimize_room_impl with probe arrivals should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_impl_progress_callback_receives_updates() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let call_count = Arc::new(AtomicUsize::new(0));
-    let count_clone = Arc::clone(&call_count);
-
-    let callback: RoomOptimizationCallback =
-        Box::new(move |_progress: &RoomOptimizationProgress| {
-            count_clone.fetch_add(1, Ordering::SeqCst);
-            CallbackAction::Continue
-        });
-    let observer = callback_pipeline_observer(callback);
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        Some(observer),
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "optimize_room_impl with progress callback should succeed: {:?}",
-        result.err()
-    );
-    assert!(
-        call_count.load(Ordering::SeqCst) > 0,
-        "progress callback should have been invoked"
-    );
-}
-
-#[test]
-fn optimize_room_impl_pipeline_observer_stop_halts() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let observer = Box::new(|_event: &PipelineEvent| -> PipelineControl { PipelineControl::Stop });
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        Some(observer),
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_err(),
-        "optimize_room_impl should halt when observer requests stop"
-    );
-}
-
-// =============================================================================
-// Additional branch coverage for optimize_room_impl and decomposed helpers
-// =============================================================================
-
-fn optimizer_for_mode(processing_mode: ProcessingMode) -> OptimizerConfig {
-    OptimizerConfig {
-        processing_mode,
-        num_filters: 1,
-        max_iter: 20,
-        population: 6,
-        seed: Some(1),
-        min_freq: 20.0,
-        max_freq: 500.0,
-        psychoacoustic: false,
-        refine: false,
-        fir: Some(crate::roomeq::types::FirConfig {
-            taps: 128,
-            phase: "linear".to_string(),
-            correct_excess_phase: false,
-            phase_smoothing: 1.0 / 6.0,
-            pre_ringing: None,
-        }),
-        mixed_phase: Some(crate::roomeq::types::MixedPhaseSerdeConfig {
-            max_fir_length_ms: 5.0,
-            pre_ringing_threshold_db: -30.0,
-            min_spatial_depth: 0.5,
-            phase_smoothing_octaves: 1.0 / 6.0,
-        }),
-        ..OptimizerConfig::default()
-    }
-}
-
-fn stereo_config_for_mode(processing_mode: ProcessingMode) -> RoomConfig {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-
-    RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: Some(SystemConfig {
-            model: SystemModel::Stereo,
-            speakers: HashMap::from([
-                ("Left".to_string(), "left".to_string()),
-                ("Right".to_string(), "right".to_string()),
-            ]),
-            subwoofers: None,
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: None,
-        target_curve: None,
-        optimizer: optimizer_for_mode(processing_mode),
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    }
-}
-
-#[test]
-fn optimize_room_impl_workflow_phase_linear_succeeds() {
-    let config = stereo_config_for_mode(ProcessingMode::PhaseLinear);
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "phase-linear stereo workflow should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.channels.is_empty());
-}
-
-#[test]
-fn optimize_room_impl_workflow_hybrid_succeeds() {
-    let config = stereo_config_for_mode(ProcessingMode::Hybrid);
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "hybrid stereo workflow should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.channels.is_empty());
-}
-
-#[test]
-fn optimize_room_impl_workflow_mixed_phase_succeeds() {
-    let config = stereo_config_for_mode(ProcessingMode::MixedPhase);
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "mixed-phase stereo workflow should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.channels.is_empty());
-}
-
-#[test]
-fn optimize_room_impl_workflow_phase_correction_succeeds() {
-    let mut config = stereo_config_for_mode(ProcessingMode::LowLatency);
-    config.optimizer.phase_correction = Some(crate::roomeq::types::MixedPhaseSerdeConfig {
-        max_fir_length_ms: 5.0,
-        pre_ringing_threshold_db: -30.0,
-        min_spatial_depth: 0.5,
-        phase_smoothing_octaves: 1.0 / 6.0,
-    });
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "workflow with phase correction should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_impl_generic_multiple_channels_phase_linear_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let mut optimizer = optimizer_for_mode(ProcessingMode::PhaseLinear);
-    optimizer.fir = Some(crate::roomeq::types::FirConfig {
-        taps: 128,
-        phase: "linear".to_string(),
-        correct_excess_phase: false,
-        phase_smoothing: 1.0 / 6.0,
-        pre_ringing: None,
-    });
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "generic multi-channel phase-linear optimization should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(result.channels.len() >= 2);
-}
-
-#[test]
-fn optimize_room_impl_home_cinema_workflow_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "center".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let system = SystemConfig {
-        model: SystemModel::HomeCinema,
-        speakers: HashMap::from([
-            ("Left".to_string(), "left".to_string()),
-            ("Right".to_string(), "right".to_string()),
-            ("Center".to_string(), "center".to_string()),
-        ]),
-        subwoofers: None,
-        bass_management: None,
-        ..Default::default()
-    };
-    let config = room_config_with_optimizer(speakers, Some(system), tiny_optimizer());
-
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "home cinema workflow should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.channels.is_empty());
-}
-
-#[test]
-fn optimize_speaker_single_channel_succeeds() {
-    let source = SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve()));
-    let result = optimize_speaker("left", &source, &tiny_optimizer(), None, 48000.0, None);
-    assert!(
-        result.is_ok(),
-        "optimize_speaker should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.biquads.is_empty() || result.fir_coeffs.is_some());
-}
-
-#[test]
-fn validate_room_optimization_single_speaker_succeeds() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer = observer_none();
-    let result = validate_room_optimization(&config, &observer);
-    assert!(
-        result.is_ok(),
-        "single-speaker config should validate: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn prepare_room_optimization_with_observer_emits_event() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let event_count = Arc::new(AtomicUsize::new(0));
-    let count_clone = Arc::clone(&event_count);
-
-    let observer = Box::new(move |_event: &PipelineEvent| -> PipelineControl {
-        count_clone.fetch_add(1, Ordering::SeqCst);
-        PipelineControl::Continue
-    });
-
-    let (_shared, prepared) = prepare_room_optimization(&config, Some(observer)).unwrap();
-    assert!(prepared.speakers.contains_key("left"));
-}
-
-// =============================================================================
-// Additional coverage for top-level entry points and observer error branches
-// =============================================================================
-
-fn stop_on_observer(
-    step: PipelineStepId,
-    status: PipelineStepStatus,
-) -> Option<Box<dyn PipelineObserver>> {
-    Some(Box::new(move |event: &PipelineEvent| {
-        if event.step_id == step && event.status == status {
-            PipelineControl::Stop
-        } else {
-            PipelineControl::Continue
-        }
-    }))
-}
-
-fn shared_stop_on_observer(
-    step: PipelineStepId,
-    status: PipelineStepStatus,
-) -> SharedPipelineObserver {
-    Arc::new(Mutex::new(stop_on_observer(step, status)))
-}
-
-fn counting_observer() -> (SharedPipelineObserver, Arc<AtomicUsize>) {
-    let count = Arc::new(AtomicUsize::new(0));
-    let c = Arc::clone(&count);
-    let observer = Arc::new(Mutex::new(Some(Box::new(move |_event: &PipelineEvent| {
-        c.fetch_add(1, Ordering::SeqCst);
-        PipelineControl::Continue
-    }) as Box<dyn PipelineObserver>)));
-    (observer, count)
-}
-
-#[test]
-fn optimize_room_with_probe_arrivals_stereo_config() {
-    let mut config = stereo_2_0_config();
-    config.optimizer = tiny_optimizer();
-    let mut probe = HashMap::new();
-    probe.insert("left".to_string(), 2.5);
-    probe.insert("right".to_string(), 3.0);
-    let result = optimize_room_with_probe_arrivals(&config, 48000.0, None, None, &probe);
-    assert!(
-        result.is_ok(),
-        "stereo config with probe arrivals should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_pipeline_impl_direct_call() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let request = RoomPipelineRequest {
-        config: &config,
-        sample_rate: 48000.0,
-        output_dir: None,
-        probe_arrival_overrides: None,
-    };
-    let result = optimize_room_pipeline_impl(request, None);
-    assert!(
-        result.is_ok(),
-        "pipeline impl direct call should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn prepare_room_optimization_observer_stop_on_started() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer = stop_on_observer(
-        PipelineStepId::ConfigPreparation,
-        PipelineStepStatus::Started,
-    );
-    let result = prepare_room_optimization(&config, observer);
-    assert!(
-        result.is_err(),
-        "observer stop on preparation started should error"
-    );
-}
-
-#[test]
-fn prepare_room_optimization_observer_stop_on_completed() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer = stop_on_observer(
-        PipelineStepId::ConfigPreparation,
-        PipelineStepStatus::Completed,
-    );
-    let result = prepare_room_optimization(&config, observer);
-    assert!(
-        result.is_err(),
-        "observer stop on preparation completed should error"
-    );
-}
-
-#[test]
-fn validate_room_optimization_observer_stop_on_started() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer = shared_stop_on_observer(PipelineStepId::Validation, PipelineStepStatus::Started);
-    let result = validate_room_optimization(&config, &observer);
-    assert!(
-        result.is_err(),
-        "observer stop on validation started should error"
-    );
-}
-
-#[test]
-fn validate_room_optimization_observer_stop_on_completed() {
-    let config = minimal_room_config(ProcessingMode::LowLatency);
-    let observer =
-        shared_stop_on_observer(PipelineStepId::Validation, PipelineStepStatus::Completed);
-    let result = validate_room_optimization(&config, &observer);
-    assert!(
-        result.is_err(),
-        "observer stop on validation completed should error"
-    );
-}
-
-#[test]
-fn select_topology_route_observer_emits_events() {
-    let config = stereo_2_0_config();
-    let (observer, count) = counting_observer();
-    let route = select_topology_route(&config, &observer).unwrap();
-    assert_eq!(route, TopologyRoute::Stereo2_0);
-    assert!(
-        count.load(Ordering::SeqCst) >= 2,
-        "topology route selection should emit started and completed events"
-    );
-}
-
-#[test]
-fn select_topology_route_observer_stop_on_started() {
-    let config = stereo_2_0_config();
-    let observer = shared_stop_on_observer(
-        PipelineStepId::TopologyRouteSelection,
-        PipelineStepStatus::Started,
-    );
-    let result = select_topology_route(&config, &observer);
-    assert!(
-        result.is_err(),
-        "observer stop on route selection started should error"
-    );
-}
-
-#[test]
-fn execute_generic_channels_multiple_speakers() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let (generic, total) =
-        execute_generic_channels(&config, 48000.0, None, None, &observer_none()).unwrap();
-    assert_eq!(total, 2);
-    assert!(generic.channel_results.contains_key("left"));
-    assert!(generic.channel_results.contains_key("right"));
-}
-
-#[test]
-fn execute_generic_channels_with_probe_arrivals() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    let mut probe = HashMap::new();
-    probe.insert("left".to_string(), 1.0);
-    probe.insert("right".to_string(), 2.0);
-    let (_generic, total) =
-        execute_generic_channels(&config, 48000.0, None, Some(&probe), &observer_none()).unwrap();
-    assert_eq!(total, 2);
-}
-
-#[test]
-fn execute_generic_channels_with_group_speaker() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "group".to_string(),
-        SpeakerConfig::Group(SpeakerGroup {
-            name: "group".to_string(),
-            speaker_name: None,
-            measurements: vec![
-                MeasurementSource::InMemory(flat_curve()),
-                MeasurementSource::InMemory(flat_curve()),
-            ],
-            crossover: Some("xo".to_string()),
-        }),
-    );
-    let mut crossovers = HashMap::new();
-    crossovers.insert(
-        "xo".to_string(),
-        CrossoverConfig {
-            crossover_type: "LR24".to_string(),
-            frequency: Some(80.0),
-            frequencies: None,
-            frequency_range: None,
-        },
-    );
-    let mut config = room_config_with_optimizer(speakers, None, tiny_optimizer());
-    config.crossovers = Some(crossovers);
-    let (generic, total) =
-        execute_generic_channels(&config, 48000.0, None, None, &observer_none()).unwrap();
-    assert_eq!(total, 1);
-    assert!(generic.channel_results.contains_key("group"));
-}
-
-#[test]
-fn assemble_workflow_result_observer_stop_on_summary() {
-    let config = stereo_2_0_config();
-    let sys = config.system.as_ref().unwrap();
-    let result = RoomOptimizationResult {
-        channels: HashMap::new(),
-        channel_results: HashMap::new(),
-        combined_pre_score: 0.0,
-        combined_post_score: 0.0,
-        metadata: empty_metadata(),
-    };
-    let observer = shared_stop_on_observer(
-        PipelineStepId::TopologyWorkflowExecution,
-        PipelineStepStatus::Completed,
-    );
-    let assembled = assemble_workflow_result(
-        result,
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        assembled.is_err(),
-        "observer stop on workflow summary should error"
-    );
-}
-
-fn two_channel_generic_collection() -> GenericChannelCollection {
-    let left = "left".to_string();
-    let right = "right".to_string();
-    let curve = flat_curve();
-    let chain = |name: &str| ChannelDspChain {
-        channel: name.to_string(),
-        plugins: Vec::new(),
-        drivers: None,
-        initial_curve: None,
-        final_curve: None,
-        eq_response: None,
-        target_curve: None,
-        pre_ir: None,
-        post_ir: None,
-        fir_temporal_masking: None,
-        direct_early_late_correction: None,
-    };
-    let mut channel_chains = HashMap::new();
-    channel_chains.insert(left.clone(), chain(&left));
-    channel_chains.insert(right.clone(), chain(&right));
-    let mut channel_results = HashMap::new();
-    channel_results.insert(
-        left.clone(),
-        ChannelOptimizationResult {
-            name: left.clone(),
-            pre_score: 0.5,
-            post_score: 0.9,
-            initial_curve: curve.clone(),
-            final_curve: curve.clone(),
-            biquads: Vec::new(),
-            fir_coeffs: None,
-        },
-    );
-    channel_results.insert(
-        right.clone(),
-        ChannelOptimizationResult {
-            name: right.clone(),
-            pre_score: 0.4,
-            post_score: 0.8,
-            initial_curve: curve.clone(),
-            final_curve: curve.clone(),
-            biquads: Vec::new(),
-            fir_coeffs: None,
-        },
-    );
-    let mut curves = HashMap::new();
-    curves.insert(left.clone(), curve.clone());
-    curves.insert(right.clone(), curve.clone());
-    let mut channel_means = HashMap::new();
-    channel_means.insert(left.clone(), 80.0);
-    channel_means.insert(right.clone(), 80.0);
-    let mut channel_arrivals = HashMap::new();
-    channel_arrivals.insert(left.clone(), 0.0);
-    channel_arrivals.insert(right.clone(), 1.0);
-    GenericChannelCollection {
-        channel_chains,
-        channel_results,
-        pre_scores: vec![0.5, 0.4],
-        post_scores: vec![0.9, 0.8],
-        curves,
-        channel_means,
-        channel_arrivals,
-    }
-}
-
-#[test]
-fn assemble_generic_result_with_observer_emits_events() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let mut optimizer = optimizer_for_mode(ProcessingMode::PhaseLinear);
-    optimizer.allow_delay = Some(true);
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-    let generic = two_channel_generic_collection();
-    let (observer, count) = counting_observer();
-    let result = assemble_generic_result(
-        generic,
-        2,
-        &config,
-        48000.0,
-        None,
-        &observer,
-        &crate::MemoryArtifactStore::new(),
-    )
-    .unwrap();
-    assert!(result.channel_results.contains_key("left"));
-    assert!(result.channel_results.contains_key("right"));
-    assert!(
-        count.load(Ordering::SeqCst) > 0,
-        "observer should receive events"
-    );
-}
-
-#[test]
-fn assemble_generic_result_multiple_channels_time_alignment() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let mut optimizer = optimizer_for_mode(ProcessingMode::PhaseLinear);
-    optimizer.allow_delay = Some(true);
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-    let generic = two_channel_generic_collection();
-    let result = assemble_generic_result(
-        generic,
-        2,
-        &config,
-        48000.0,
-        None,
-        &observer_none(),
-        &crate::MemoryArtifactStore::new(),
-    )
-    .unwrap();
-    assert!(result.channels.contains_key("left"));
-    assert!(result.channels.contains_key("right"));
-}
-
-#[test]
-fn optimize_speaker_group_missing_crossover_fails() {
-    let group = SpeakerConfig::Group(SpeakerGroup {
-        name: "group".to_string(),
-        speaker_name: None,
-        measurements: vec![MeasurementSource::InMemory(flat_curve())],
-        crossover: None,
-    });
-    let result = optimize_speaker("group", &group, &tiny_optimizer(), None, 48000.0, None);
-    assert!(
-        result.is_err(),
-        "speaker group without crossover should fail: {:?}",
-        result
-    );
-}
-
-#[test]
-fn optimize_speaker_with_target_curve_succeeds() {
-    let source = SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve()));
-    let target = Some(TargetCurveConfig::Predefined("flat".into()));
-    let result = optimize_speaker(
-        "left",
-        &source,
-        &tiny_optimizer(),
-        target.as_ref(),
-        48000.0,
-        None,
-    );
-    assert!(
-        result.is_ok(),
-        "optimize_speaker with target curve should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_speaker_group_no_measurements_fails() {
-    let group = SpeakerConfig::Group(SpeakerGroup {
-        name: "group".to_string(),
-        speaker_name: None,
-        measurements: vec![],
-        crossover: None,
-    });
-    let result = optimize_speaker("group", &group, &tiny_optimizer(), None, 48000.0, None);
-    assert!(result.is_err(), "group with no measurements should fail");
-}
-
-#[test]
-fn optimize_room_impl_stereo_2_1_workflow_succeeds() {
-    let config = stereo_2_1_config();
-    let result = optimize_room_impl(
-        &config,
-        48000.0,
-        None,
-        None,
-        None,
-        &crate::MemoryArtifactStore::new(),
-    );
-    assert!(
-        result.is_ok(),
-        "stereo 2.1 workflow should succeed: {:?}",
-        result.err()
-    );
-    let result = result.unwrap();
-    assert!(!result.channels.is_empty());
-}
-
-#[test]
-fn execute_topology_workflow_home_cinema_with_sub_returns_result() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "center".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "sub".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let mut crossovers = HashMap::new();
-    crossovers.insert(
-        "xo".to_string(),
-        CrossoverConfig {
-            crossover_type: "LR24".to_string(),
-            frequency: Some(80.0),
-            frequencies: None,
-            frequency_range: None,
-        },
-    );
-    let config = RoomConfig {
-        version: crate::roomeq::types::default_config_version(),
-        system: Some(SystemConfig {
-            model: SystemModel::HomeCinema,
-            speakers: HashMap::from([
-                ("Left".to_string(), "left".to_string()),
-                ("Right".to_string(), "right".to_string()),
-                ("Center".to_string(), "center".to_string()),
-                ("LFE".to_string(), "sub".to_string()),
-            ]),
-            subwoofers: Some(SubwooferSystemConfig {
-                config: SubwooferStrategy::Single,
-                crossover: Some("xo".to_string()),
-                mapping: [("sub".to_string(), "Center".to_string())].into(),
-            }),
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: Some(crossovers),
-        target_curve: None,
-        optimizer: tiny_optimizer(),
-        recording_config: None,
-        ctc: None,
-        cea2034_cache: None,
-    };
-    let sys = config.system.as_ref().unwrap();
-    let result = execute_topology_workflow(
-        &config,
-        sys,
-        48000.0,
-        None,
-        &observer_none(),
-        TopologyRoute::HomeCinema,
-    )
-    .unwrap();
-    assert!(
-        !result.channels.is_empty(),
-        "home cinema with sub should produce channels"
-    );
-}
-
-// Additional coverage tests for high-level optimize_room branches
-
-#[test]
-fn optimize_room_with_group_delay_enabled_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.optimizer.group_delay = Some(crate::roomeq::types::GroupDelayOptimizationConfig {
-        enabled: true,
-        max_iter: 20,
-        popsize: 4,
-        tol: 1e-3,
-        min_improvement_db: 0.0,
-        ..Default::default()
-    });
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with GD enabled should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_phase_alignment_enabled_succeeds() {
-    let mut config = stereo_2_1_config();
-    config.optimizer.allow_delay = Some(true);
-    config.optimizer.phase_alignment = Some(crate::roomeq::types::PhaseAlignmentConfig {
-        enabled: true,
-        ..Default::default()
-    });
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with phase alignment should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_vog_enabled_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.optimizer.vog = Some(crate::roomeq::types::VoiceOfGodConfig {
-        enabled: true,
-        reference_channel: "Left".to_string(),
-    });
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with VoG should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_target_curve_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.target_curve = Some(crate::roomeq::types::TargetCurveConfig::Predefined(
-        "flat".to_string(),
-    ));
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with target curve should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_psychoacoustic_smoothing_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.optimizer.psychoacoustic = true;
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with psychoacoustic smoothing should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_refine_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.optimizer.refine = true;
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with refine should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_cobyla_algorithm_succeeds() {
-    let mut config = stereo_2_0_config();
-    config.optimizer.algorithm = "autoeq:cobyla".to_string();
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with cobyla should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_generic_with_time_alignment_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    speakers.insert(
-        "right".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemory(flat_curve())),
-    );
-    let mut optimizer = tiny_optimizer();
-    optimizer.allow_delay = Some(true);
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "generic optimize_room with time alignment should succeed: {:?}",
-        result.err()
-    );
-}
-
-#[test]
-fn optimize_room_with_multi_measurement_weighted_succeeds() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "left".to_string(),
-        SpeakerConfig::Single(MeasurementSource::InMemoryMultiple(vec![
-            flat_curve(),
-            flat_curve(),
-        ])),
-    );
-    let mut optimizer = tiny_optimizer();
-    optimizer.multi_measurement = Some(MultiMeasurementConfig {
-        strategy: MultiMeasurementStrategy::WeightedSum,
-        weights: Some(vec![0.6, 0.4]),
-        ..Default::default()
-    });
-    let config = room_config_with_optimizer(speakers, None, optimizer);
-    let result = optimize_room(&config, 48000.0, None, None);
-    assert!(
-        result.is_ok(),
-        "optimize_room with multi-measurement weighted should succeed: {:?}",
-        result.err()
-    );
-}
+include!("tests/topology.rs");
+include!("tests/pipeline.rs");
+include!("tests/processing_modes.rs");
+include!("tests/observers.rs");
+include!("tests/workflows.rs");
