+++ src/roomeq/types/config/tests.rs
@@ -34,1246 +34,6 @@ use crate::{MeasurementRef, MeasurementSingle, MeasurementSource};
 use std::collections::HashMap;
 use std::path::PathBuf;
 
-#[test]
-fn spl_calibration_roundtrip_and_helpers() {
-    let cal = SplCalibration {
-        reported_db_spl: 85.0,
-        reference_freq_hz: 1000.0,
-        peak_sample_level: 0.5,
-        spl_offset_db: 85.0 - 20.0 * 0.5_f32.log10(),
-    };
-    // Round-trip through JSON.
-    let json = serde_json::to_string(&cal).unwrap();
-    let back: SplCalibration = serde_json::from_str(&json).unwrap();
-    assert_eq!(back, cal);
-    // `dbspl_for_peak_level` at the calibration peak must return the
-    // reported dBSPL (within f32 rounding).
-    let recovered = cal.dbspl_for_peak_level(cal.peak_sample_level);
-    assert!((recovered - cal.reported_db_spl).abs() < 1e-3);
-    // `peak_level_for_dbspl` is the inverse.
-    let peak = cal.peak_level_for_dbspl(cal.reported_db_spl);
-    assert!((peak - cal.peak_sample_level).abs() < 1e-5);
-    // Clamp: asking for a dBSPL below the representable range
-    // should return 0.0 rather than a negative peak level.
-    assert_eq!(cal.peak_level_for_dbspl(-1000.0), 0.0);
-}
-
-#[test]
-fn recording_configuration_accepts_gd_v2_fields() {
-    let cfg_json = serde_json::json!({
-        "bass_octave_duration_s": 3.0,
-        "pre_silence_s": 2.0,
-        "post_silence_s": 4.0,
-        "sweep_level_db_spl": 85.0,
-        "num_sweeps": 4,
-        "coherence_threshold": 0.9,
-        "bass_probe_freq_hz": 30.0,
-        "bass_probe_duration_s": 2.0,
-        "mic_phase_calibration_path": "/tmp/mic_phase.csv",
-        "spl_calibration": {
-            "reported_db_spl": 85.0,
-            "reference_freq_hz": 1000.0,
-            "peak_sample_level": 0.5,
-            "spl_offset_db": 91.02
-        },
-        "recording_seed": 42
-    });
-    let cfg: RecordingConfiguration = serde_json::from_value(cfg_json).unwrap();
-    assert_eq!(cfg.bass_octave_duration_s, Some(3.0));
-    assert_eq!(cfg.num_sweeps, Some(4));
-    assert_eq!(cfg.coherence_threshold, Some(0.9));
-    assert_eq!(cfg.bass_probe_freq_hz, Some(30.0));
-    assert_eq!(cfg.bass_probe_duration_s, Some(2.0));
-    assert_eq!(
-        cfg.mic_phase_calibration_path.as_deref(),
-        Some("/tmp/mic_phase.csv")
-    );
-    let cal = cfg.spl_calibration.expect("spl_calibration populated");
-    assert!((cal.reported_db_spl - 85.0).abs() < 1e-6);
-    assert_eq!(cfg.recording_seed, Some(42));
-}
-
-#[test]
-fn recording_configuration_legacy_json_still_loads() {
-    // A session written before GD-1a only knows the pre-existing
-    // fields. GD-Opt v2 metadata must default to `None` rather
-    // than failing deserialization — the confidence gate is what
-    // converts `None` into `Advisory::GdOptDegradedPhase` at
-    // optimization time.
-    let legacy_json = serde_json::json!({
-        "signal_type": "Sweep",
-        "signal_duration_secs": 10.0,
-        "sweep_start_freq": 20.0,
-        "sweep_end_freq": 20000.0
-    });
-    let cfg: RecordingConfiguration = serde_json::from_value(legacy_json).unwrap();
-    assert!(cfg.bass_octave_duration_s.is_none());
-    assert!(cfg.num_sweeps.is_none());
-    assert!(cfg.coherence_threshold.is_none());
-    assert!(cfg.bass_probe_freq_hz.is_none());
-    assert!(cfg.bass_probe_duration_s.is_none());
-    assert!(cfg.mic_phase_calibration_path.is_none());
-    assert!(cfg.mic_phase_calibration_paths.is_none());
-    assert!(cfg.spl_calibration.is_none());
-    assert!(cfg.recording_seed.is_none());
-}
-
-#[test]
-fn bass_anchor_results_legacy_migrates_v1_bass_cycles_to_duration() {
-    // Pre-v2 schema: bass_cycles + bass_freq_hz, no bass_duration_s.
-    let legacy = serde_json::json!({
-        "channels": [],
-        "sample_rate": 48_000_u32,
-        "bass_freq_hz": 30.0_f32,
-        "bass_cycles": 6_u16,
-    });
-    let r: BassAnchorResultsLegacy = serde_json::from_value(legacy).unwrap();
-    // 6 cycles at 30 Hz = 0.2 s.
-    assert!(
-        (r.bass_duration_s - 0.2).abs() < 1e-6,
-        "expected 0.2 s migrated from 6 cycles @ 30 Hz, got {}",
-        r.bass_duration_s
-    );
-    assert_eq!(r.sample_rate, 48_000);
-    assert_eq!(r.bass_freq_hz, 30.0);
-}
-
-#[test]
-fn bass_anchor_results_prefers_explicit_duration_when_both_present() {
-    let mixed = serde_json::json!({
-        "channels": [],
-        "sample_rate": 48_000_u32,
-        "bass_freq_hz": 30.0_f32,
-        "bass_cycles": 6_u16,
-        "bass_duration_s": 2.5_f32,
-    });
-    let r: BassAnchorResultsLegacy = serde_json::from_value(mixed).unwrap();
-    assert!((r.bass_duration_s - 2.5).abs() < 1e-6);
-}
-
-#[test]
-fn bass_anchor_results_v2_round_trips() {
-    let v2 = serde_json::json!({
-        "channels": [],
-        "sample_rate": 48_000_u32,
-        "bass_freq_hz": 30.0_f32,
-        "bass_duration_s": 2.0_f32,
-    });
-    let r: BassAnchorResultsLegacy = serde_json::from_value(v2).unwrap();
-    assert_eq!(r.bass_duration_s, 2.0);
-}
-
-#[test]
-fn target_shape_canonical_wire_format() {
-    // Pins the on-the-wire string for every TargetShape variant.
-    // `from_measurement` (underscore) is the sole canonical form,
-    // matching bin/roomeq/input_schema.json and INPUT_FORMAT.md.
-    let cases = [
-        (TargetShape::Flat, "\"flat\""),
-        (TargetShape::Harman, "\"harman\""),
-        (TargetShape::Custom, "\"custom\""),
-        (TargetShape::File, "\"file\""),
-        (TargetShape::FromMeasurement, "\"from_measurement\""),
-    ];
-    for (variant, expected) in cases {
-        let serialized = serde_json::to_string(&variant).unwrap();
-        assert_eq!(serialized, expected, "serialize {variant:?}");
-        let round_tripped: TargetShape = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(round_tripped, variant, "round-trip {variant:?}");
-    }
-    // Old canonical form used before the snake_case switch must no
-    // longer deserialize — a paranoid guard against accidental
-    // reintroduction of the `#[serde(alias = "from_measurement")]`
-    // back-compat shim.
-    assert!(serde_json::from_str::<TargetShape>("\"frommeasurement\"").is_err());
-}
-
-#[test]
-fn test_optimizer_config_default_has_decomposed_correction_enabled() {
-    let config = OptimizerConfig::default();
-    let dc = config
-        .decomposed_correction
-        .expect("decomposed_correction should be Some by default");
-    assert!(
-        dc.enabled,
-        "decomposed_correction should be enabled by default"
-    );
-    assert_eq!(dc.schroeder_freq, 250.0);
-    assert_eq!(dc.steady_state_weight, 0.4);
-}
-
-#[test]
-fn test_optimizer_config_default_algorithm_is_cmaes() {
-    let config = OptimizerConfig::default();
-    assert_eq!(config.algorithm, "autoeq:cmaes");
-}
-
-#[test]
-fn test_optimizer_config_resolves_new_acoustic_defaults() {
-    let config = OptimizerConfig::default();
-    assert_eq!(
-        config.psychoacoustic_smoothing_config(),
-        PsychoacousticSmoothingConfig::default()
-    );
-    assert_eq!(
-        config.asymmetric_loss_config(),
-        AsymmetricLossConfig::default()
-    );
-
-    let custom_smoothing = PsychoacousticSmoothingConfig {
-        low_freq_n: 24,
-        high_freq_n: 3,
-        low_freq: 80.0,
-        high_freq: 800.0,
-    };
-    let custom_asym = AsymmetricLossConfig {
-        peak_weight: 3.0,
-        dip_weight: 0.5,
-        bass_peak_weight: 6.0,
-        bass_dip_weight: 0.25,
-        transition_freq: 180.0,
-    };
-    let config = OptimizerConfig {
-        psychoacoustic_smoothing: Some(custom_smoothing),
-        asymmetric_loss_config: Some(custom_asym),
-        ..Default::default()
-    };
-    assert_eq!(config.psychoacoustic_smoothing_config(), custom_smoothing);
-    assert_eq!(config.asymmetric_loss_config(), custom_asym);
-}
-
-#[test]
-fn perceptual_policy_preserves_legacy_defaults_when_absent() {
-    let mut config = OptimizerConfig::default();
-    config.apply_perceptual_policy_defaults();
-
-    assert!(config.perceptual_policy.is_none());
-    assert!(config.audibility_deadband.is_none());
-    assert!(config.high_frequency_correction.is_none());
-    assert!(config.early_late_correction.is_none());
-    assert!(config.validation_bundle.is_none());
-    assert_eq!(config.loss_type, "flat");
-    assert_eq!(config.max_freq, default_max_freq());
-}
-
-#[test]
-fn music_policy_fills_perceptual_defaults() {
-    let mut config = OptimizerConfig {
-        perceptual_policy: Some(PerceptualPolicyConfig {
-            preset: PerceptualPolicyPreset::Music,
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-
-    config.apply_perceptual_policy_defaults();
-
-    assert_eq!(config.loss_type, "epa");
-    assert_eq!(
-        config.target_response.as_ref().unwrap().shape,
-        TargetShape::Harman
-    );
-    assert!(config.audibility_deadband_config().is_some());
-    assert!(config.high_frequency_correction.is_some());
-    assert!(config.early_late_correction_config().is_some());
-    assert!(config.validation_bundle_config().is_some());
-    assert_eq!(
-        config.multi_measurement.as_ref().unwrap().strategy,
-        MultiMeasurementStrategy::SpatialRobustness
-    );
-}
-
-#[test]
-fn high_frequency_guard_caps_default_q_but_preserves_explicit_q() {
-    let mut default_q = OptimizerConfig {
-        max_freq: 8_000.0,
-        high_frequency_correction: Some(HighFrequencyCorrectionConfig::default()),
-        ..Default::default()
-    };
-    default_q.apply_high_frequency_correction_defaults(false);
-    assert_eq!(default_q.max_q, default_high_freq_guard_max_q());
-    assert_eq!(
-        default_q.psychoacoustic_smoothing.unwrap().high_freq_n,
-        default_high_freq_smoothing_n()
-    );
-
-    let mut explicit_q = OptimizerConfig {
-        max_freq: 8_000.0,
-        max_q: 4.5,
-        high_frequency_correction: Some(HighFrequencyCorrectionConfig::default()),
-        ..Default::default()
-    };
-    explicit_q.apply_high_frequency_correction_defaults(false);
-    assert_eq!(explicit_q.max_q, 4.5);
-
-    explicit_q.apply_high_frequency_correction_defaults(true);
-    assert_eq!(explicit_q.max_q, default_high_freq_guard_max_q());
-}
-
-#[test]
-fn perceptual_policy_tightens_generated_high_frequency_smoothing() {
-    let mut config = OptimizerConfig {
-        max_freq: 8_000.0,
-        perceptual_policy: Some(PerceptualPolicyConfig {
-            preset: PerceptualPolicyPreset::Music,
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-
-    config.apply_perceptual_policy_defaults();
-
-    let smoothing = config.psychoacoustic_smoothing.unwrap();
-    assert_eq!(
-        smoothing.high_freq_n,
-        policy_high_frequency_guard(PerceptualPolicyPreset::Music).smoothing_n
-    );
-    assert!(smoothing.high_freq <= default_high_freq_guard_start_hz());
-}
-
-#[test]
-fn test_decomposed_correction_serde_config_default() {
-    let dc = DecomposedCorrectionSerdeConfig::default();
-    assert!(dc.enabled);
-    assert_eq!(dc.schroeder_freq, 250.0);
-    assert_eq!(dc.steady_state_weight, 0.4);
-    assert_eq!(dc.min_mode_q, 3.0);
-    assert_eq!(dc.min_mode_prominence_db, 3.0);
-    assert_eq!(dc.mode_correction_weight, 1.0);
-    assert_eq!(dc.early_reflection_weight, 0.3);
-    assert!(dc.fdw_enabled);
-    assert_eq!(dc.fdw_cycles, 8.0);
-    assert_eq!(dc.fdw_min_window_ms, 3.0);
-    assert_eq!(dc.fdw_max_window_ms, 500.0);
-    assert_eq!(dc.fdw_smoothing_octaves, 1.0 / 24.0);
-}
-
-#[test]
-fn test_channel_matching_config_defaults() {
-    let cfg = ChannelMatchingConfig::default();
-    assert!(cfg.enabled);
-    assert_eq!(cfg.threshold_db, 0.75);
-    assert_eq!(cfg.max_filters, 5);
-}
-
-#[test]
-fn test_max_boost_envelope_interpolation() {
-    let mut config = OptimizerConfig::default();
-
-    // Without envelope, falls back to max_db
-    assert_eq!(config.max_boost_at_freq(100.0), config.max_db);
-
-    // Set an envelope: generous bass boost tapering to zero
-    config.max_boost_envelope = Some(vec![
-        (20.0, 6.0),
-        (200.0, 4.0),
-        (1000.0, 2.0),
-        (8000.0, 0.0),
-    ]);
-
-    // At exact envelope points
-    assert!((config.max_boost_at_freq(20.0) - 6.0).abs() < 1e-10);
-    assert!((config.max_boost_at_freq(200.0) - 4.0).abs() < 1e-10);
-    assert!((config.max_boost_at_freq(1000.0) - 2.0).abs() < 1e-10);
-    assert!((config.max_boost_at_freq(8000.0) - 0.0).abs() < 1e-10);
-
-    // Below first point: clamp to first value
-    assert!((config.max_boost_at_freq(10.0) - 6.0).abs() < 1e-10);
-
-    // Above last point: clamp to last value
-    assert!((config.max_boost_at_freq(16000.0) - 0.0).abs() < 1e-10);
-
-    // Between 200Hz and 1000Hz: log-frequency interpolation
-    // Geometric midpoint of 200 and 1000 is sqrt(200*1000) ~ 447Hz
-    let mid_freq = (200.0_f64 * 1000.0).sqrt();
-    let mid_boost = config.max_boost_at_freq(mid_freq);
-    // At geometric midpoint, t = 0.5, so interpolated value = 4.0 + 0.5*(2.0-4.0) = 3.0
-    assert!(
-        (mid_boost - 3.0).abs() < 1e-6,
-        "geometric midpoint should give 3.0 dB, got {:.6}",
-        mid_boost
-    );
-}
-
-#[test]
-fn config_types_default_roundtrip() {
-    use super::*;
-
-    let _ = AreaPriorKind::default();
-    let _ = AreaQuadratureKind::default();
-    let _ = AreaScalarisationKind::default();
-    let _ = BootstrapScalarisation::default();
-    let _ = CtcConfig::default();
-    let _ = CtcRegularizationConfig::default();
-    let _ = CtcWindowConfig::default();
-    let _ = MixedModeConfig::default();
-    let _ = SubOptimizerConfig::default();
-
-    let _ = serde_json::to_string(&CtcConfig::default()).unwrap();
-    let _ = serde_json::to_string(&MixedModeConfig::default()).unwrap();
-    let _ = serde_json::to_string(&SubOptimizerConfig::default()).unwrap();
-
-    let quad = AreaQuadratureKind::Sobol {
-        num_points: 16,
-        seed: 42,
-    };
-    let json = serde_json::to_string(&quad).unwrap();
-    let back: AreaQuadratureKind = serde_json::from_str(&json).unwrap();
-    assert_eq!(back, quad);
-}
-
-// ---------------------------------------------------------------------------
-// Tests for roomeq/types/config/* modules that previously had little or no
-// unit-test coverage.  Each test exercises construction, JSON round-tripping,
-// and path-resolution helpers where applicable.
-// ---------------------------------------------------------------------------
-
-fn sample_single_source(path: &str, speaker_name: Option<&str>) -> MeasurementSource {
-    MeasurementSource::Single(MeasurementSingle {
-        measurement: MeasurementRef::Path(path.into()),
-        speaker_name: speaker_name.map(|s| s.to_string()),
-    })
-}
-
-#[test]
-fn cardioid_config_roundtrip_and_resolve_paths() {
-    let mut cfg = CardioidConfig {
-        name: "sub-cardioid".into(),
-        speaker_name: Some("Subwoofer X".into()),
-        front: sample_single_source("front.csv", None),
-        rear: sample_single_source("rear.csv", Some("Rear Sub")),
-        separation_meters: 0.35,
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: CardioidConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.name, cfg.name);
-    assert_eq!(back.speaker_name, cfg.speaker_name);
-    assert_eq!(back.separation_meters, cfg.separation_meters);
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back.front {
-        MeasurementSource::Single(s) => match &s.measurement {
-            MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/front.csv")),
-            other => panic!("expected path front, got {:?}", other),
-        },
-        other => panic!("expected single front, got {:?}", other),
-    }
-    match &back.rear {
-        MeasurementSource::Single(s) => assert_eq!(s.speaker_name.as_deref(), Some("Rear Sub")),
-        other => panic!("expected single rear, got {:?}", other),
-    }
-    // resolve_paths mutates in place; reuse the resolved value for coverage.
-    cfg.resolve_paths(PathBuf::from("/another").as_path());
-}
-
-#[test]
-fn ctc_config_default_and_roundtrip_and_resolve_paths() {
-    let default = CtcConfig::default();
-    assert!(!default.enabled);
-    assert_eq!(default.matrix_source, "measured");
-    assert_eq!(default.robustness, "average");
-    assert!(default.include_room_eq_dsp);
-    assert_eq!(default.fir_taps, 4096);
-    assert_eq!(default.harmonic_suppression_harmonics, 5);
-    assert!(default.measurements.is_none());
-    assert!(default.hrtf.is_none());
-    assert!(default.reference_sweep.is_none());
-
-    let mut cfg = CtcConfig {
-        enabled: true,
-        measurements: Some(CtcMeasurementConfig {
-            speakers: vec!["L".into(), "R".into()],
-            mics: vec!["mic1".into()],
-            head_positions: vec![CtcHeadPositionConfig {
-                id: "center".into(),
-                x: 0.0,
-                y: 0.0,
-                z: 0.0,
-                yaw_deg: 0.0,
-            }],
-            files: vec![CtcMeasurementFileConfig {
-                head_position: "center".into(),
-                speaker: "L".into(),
-                ir: Some("ir.wav".into()),
-                raw_sweep: None,
-                loopback: None,
-            }],
-        }),
-        reference_sweep: Some("sweep.wav".into()),
-        sweep_duration_s: Some(10.0),
-        sweep_start_hz: Some(20.0),
-        sweep_end_hz: Some(20000.0),
-        ..Default::default()
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: CtcConfig = serde_json::from_str(&json).unwrap();
-    assert!(back.enabled);
-    assert!(back.measurements.is_some());
-    assert_eq!(back.measurements.as_ref().unwrap().speakers, vec!["L", "R"]);
-    assert_eq!(back.reference_sweep, Some(PathBuf::from("sweep.wav")));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    assert_eq!(back.reference_sweep, Some(PathBuf::from("/base/sweep.wav")));
-    assert_eq!(
-        back.measurements.as_ref().unwrap().files[0].ir,
-        Some(PathBuf::from("/base/ir.wav"))
-    );
-
-    // Coverage for the hrtf path-resolution branch.
-    cfg.hrtf = Some(CtcHrtfConfig {
-        hrtf_file: "hrtf.sofa".into(),
-        speakers: vec![CtcHrtfSpeakerConfig {
-            speaker: "L".into(),
-            azimuth_deg: -30.0,
-            elevation_deg: 0.0,
-            distance_m: 2.0,
-        }],
-    });
-    cfg.resolve_paths(PathBuf::from("/base").as_path());
-    assert_eq!(
-        cfg.hrtf.as_ref().unwrap().hrtf_file,
-        PathBuf::from("/base/hrtf.sofa")
-    );
-}
-
-#[test]
-fn ctc_hrtf_config_roundtrip_and_resolve_paths() {
-    let cfg = CtcHrtfConfig {
-        hrtf_file: "hrtf.sofa".into(),
-        speakers: vec![
-            CtcHrtfSpeakerConfig {
-                speaker: "L".into(),
-                azimuth_deg: -30.0,
-                elevation_deg: 0.0,
-                distance_m: 2.0,
-            },
-            CtcHrtfSpeakerConfig {
-                speaker: "R".into(),
-                azimuth_deg: 30.0,
-                elevation_deg: 0.0,
-                distance_m: 2.0,
-            },
-        ],
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: CtcHrtfConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.hrtf_file, cfg.hrtf_file);
-    assert_eq!(back.speakers.len(), 2);
-    assert_eq!(back.speakers[0].speaker, "L");
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    assert_eq!(back.hrtf_file, PathBuf::from("/base/hrtf.sofa"));
-}
-
-#[test]
-fn ctc_measurement_config_roundtrip_and_resolve_paths() {
-    let mut cfg = CtcMeasurementConfig {
-        speakers: vec!["L".into(), "R".into()],
-        mics: vec!["mic1".into(), "mic2".into()],
-        head_positions: vec![CtcHeadPositionConfig {
-            id: "center".into(),
-            x: 0.0,
-            y: 0.0,
-            z: 0.0,
-            yaw_deg: 0.0,
-        }],
-        files: vec![CtcMeasurementFileConfig {
-            head_position: "center".into(),
-            speaker: "L".into(),
-            ir: Some("ir.wav".into()),
-            raw_sweep: Some("raw.wav".into()),
-            loopback: Some("loopback.wav".into()),
-        }],
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: CtcMeasurementConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speakers, cfg.speakers);
-    assert_eq!(back.mics, cfg.mics);
-    assert_eq!(back.head_positions.len(), 1);
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    let file = &back.files[0];
-    assert_eq!(file.ir, Some(PathBuf::from("/base/ir.wav")));
-    assert_eq!(file.raw_sweep, Some(PathBuf::from("/base/raw.wav")));
-    assert_eq!(file.loopback, Some(PathBuf::from("/base/loopback.wav")));
-
-    // Absolute paths should be left untouched.
-    cfg.files[0].ir = Some("/abs/ir.wav".into());
-    cfg.resolve_paths(PathBuf::from("/base").as_path());
-    assert_eq!(cfg.files[0].ir, Some(PathBuf::from("/abs/ir.wav")));
-}
-
-#[test]
-fn dba_config_roundtrip_and_resolve_paths() {
-    let cfg = DBAConfig {
-        name: "dba".into(),
-        speaker_name: Some("Sub Array".into()),
-        front: vec![
-            sample_single_source("front0.csv", None),
-            sample_single_source("front1.csv", None),
-        ],
-        rear: vec![sample_single_source("rear0.csv", None)],
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: DBAConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.name, cfg.name);
-    assert_eq!(back.speaker_name, cfg.speaker_name);
-    assert_eq!(back.front.len(), 2);
-    assert_eq!(back.rear.len(), 1);
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back.front[0] {
-        MeasurementSource::Single(s) => match &s.measurement {
-            MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/front0.csv")),
-            other => panic!("expected path, got {:?}", other),
-        },
-        other => panic!("expected single, got {:?}", other),
-    }
-}
-
-#[test]
-fn multi_sub_group_roundtrip_and_resolve_paths() {
-    let cfg = MultiSubGroup {
-        name: "subs".into(),
-        speaker_name: Some("Sub".into()),
-        subwoofers: vec![
-            sample_single_source("sub1.csv", None),
-            sample_single_source("sub2.csv", None),
-        ],
-        allpass_optimization: true,
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: MultiSubGroup = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.name, cfg.name);
-    assert_eq!(back.speaker_name, cfg.speaker_name);
-    assert!(back.allpass_optimization);
-    assert_eq!(back.subwoofers.len(), 2);
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back.subwoofers[1] {
-        MeasurementSource::Single(s) => match &s.measurement {
-            MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/sub2.csv")),
-            other => panic!("expected path, got {:?}", other),
-        },
-        other => panic!("expected single, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_group_roundtrip_and_resolve_paths() {
-    let cfg = SpeakerGroup {
-        name: "L-speaker".into(),
-        speaker_name: Some("Model A".into()),
-        measurements: vec![
-            sample_single_source("woofer.csv", None),
-            sample_single_source("tweeter.csv", None),
-        ],
-        crossover: Some("lr24".into()),
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerGroup = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.name, cfg.name);
-    assert_eq!(back.speaker_name, cfg.speaker_name);
-    assert_eq!(back.measurements.len(), 2);
-    assert_eq!(back.crossover.as_deref(), Some("lr24"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back.measurements[0] {
-        MeasurementSource::Single(s) => match &s.measurement {
-            MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/woofer.csv")),
-            other => panic!("expected path, got {:?}", other),
-        },
-        other => panic!("expected single, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_config_single_roundtrip_and_resolve_paths() {
-    let cfg = SpeakerConfig::Single(sample_single_source("L.csv", Some("Model A")));
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speaker_name(), Some("Model A"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back {
-        SpeakerConfig::Single(s) => match s {
-            MeasurementSource::Single(ms) => match &ms.measurement {
-                MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/L.csv")),
-                other => panic!("expected path, got {:?}", other),
-            },
-            other => panic!("expected single source, got {:?}", other),
-        },
-        other => panic!("expected single variant, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_config_group_roundtrip_and_resolve_paths() {
-    let group = SpeakerGroup {
-        name: "L-speaker".into(),
-        speaker_name: Some("Model A".into()),
-        measurements: vec![sample_single_source("woofer.csv", None)],
-        crossover: None,
-    };
-    let cfg = SpeakerConfig::Group(group);
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speaker_name(), Some("Model A"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back {
-        SpeakerConfig::Group(g) => {
-            assert_eq!(g.name, "L-speaker");
-        }
-        other => panic!("expected group variant, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_config_multi_sub_roundtrip_and_resolve_paths() {
-    let multi = MultiSubGroup {
-        name: "subs".into(),
-        speaker_name: Some("Sub".into()),
-        subwoofers: vec![sample_single_source("sub.csv", None)],
-        allpass_optimization: false,
-    };
-    let cfg = SpeakerConfig::MultiSub(multi);
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speaker_name(), Some("Sub"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back {
-        SpeakerConfig::MultiSub(ms) => {
-            assert_eq!(ms.name, "subs");
-        }
-        other => panic!("expected multi-sub variant, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_config_dba_roundtrip_and_resolve_paths() {
-    let dba = DBAConfig {
-        name: "dba".into(),
-        speaker_name: Some("DBA".into()),
-        front: vec![sample_single_source("front.csv", None)],
-        rear: vec![sample_single_source("rear.csv", None)],
-    };
-    let cfg = SpeakerConfig::Dba(dba);
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speaker_name(), Some("DBA"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back {
-        SpeakerConfig::Dba(d) => {
-            assert_eq!(d.name, "dba");
-        }
-        other => panic!("expected dba variant, got {:?}", other),
-    }
-}
-
-#[test]
-fn speaker_config_cardioid_roundtrip_and_resolve_paths() {
-    let cardioid = CardioidConfig {
-        name: "cardioid".into(),
-        speaker_name: Some("Cardioid".into()),
-        front: sample_single_source("front.csv", None),
-        rear: sample_single_source("rear.csv", None),
-        separation_meters: 0.5,
-    };
-    let cfg = SpeakerConfig::Cardioid(Box::new(cardioid));
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: SpeakerConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.speaker_name(), Some("Cardioid"));
-
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back {
-        SpeakerConfig::Cardioid(c) => {
-            assert_eq!(c.name, "cardioid");
-            assert_eq!(c.separation_meters, 0.5);
-        }
-        other => panic!("expected cardioid variant, got {:?}", other),
-    }
-}
-
-#[test]
-fn room_config_roundtrip_and_resolve_paths() {
-    let mut speakers = HashMap::new();
-    speakers.insert(
-        "L".into(),
-        SpeakerConfig::Single(sample_single_source("L.csv", Some("Model A"))),
-    );
-    speakers.insert(
-        "R".into(),
-        SpeakerConfig::Single(sample_single_source("R.csv", None)),
-    );
-
-    let mut crossovers = HashMap::new();
-    crossovers.insert(
-        "L".into(),
-        CrossoverConfig {
-            crossover_type: "LR24".into(),
-            frequency: Some(2000.0),
-            frequencies: None,
-            frequency_range: None,
-        },
-    );
-
-    let mut system_speakers = HashMap::new();
-    system_speakers.insert("L".into(), "L".into());
-    system_speakers.insert("R".into(), "R".into());
-
-    let mut cfg = RoomConfig {
-        version: "1.0.0".into(),
-        system: Some(SystemConfig {
-            model: SystemModel::Stereo,
-            speakers: system_speakers,
-            subwoofers: None,
-            bass_management: None,
-            ..Default::default()
-        }),
-        speakers,
-        crossovers: Some(crossovers),
-        target_curve: Some(TargetCurveConfig::Predefined("flat".into())),
-        optimizer: OptimizerConfig::default(),
-        recording_config: None,
-        ctc: Some(CtcConfig::default()),
-        cea2034_cache: None,
-    };
-
-    let json = serde_json::to_string(&cfg).unwrap();
-    let mut back: RoomConfig = serde_json::from_str(&json).unwrap();
-    assert_eq!(back.version, cfg.version);
-    assert!(back.system.is_some());
-    assert_eq!(back.speakers.len(), 2);
-    assert!(back.crossovers.is_some());
-    assert!(back.ctc.is_some());
-    match &back.target_curve {
-        Some(TargetCurveConfig::Predefined(name)) => assert_eq!(name, "flat"),
-        other => panic!("expected predefined target curve, got {:?}", other),
-    }
-
-    // Exercise the path-resolution branch by replacing the target with a Path
-    // variant after deserialization (the untagged enum serializes strings as
-    // Predefined, so we set it explicitly here).
-    back.target_curve = Some(TargetCurveConfig::Path("target.csv".into()));
-    back.resolve_paths(PathBuf::from("/base").as_path());
-    match &back.target_curve {
-        Some(TargetCurveConfig::Path(p)) => {
-            assert_eq!(p, &PathBuf::from("/base/target.csv"))
-        }
-        other => panic!("expected path target curve, got {:?}", other),
-    }
-    match &back.speakers["L"] {
-        SpeakerConfig::Single(s) => match s {
-            MeasurementSource::Single(ms) => match &ms.measurement {
-                MeasurementRef::Path(p) => assert_eq!(p, &PathBuf::from("/base/L.csv")),
-                other => panic!("expected path, got {:?}", other),
-            },
-            other => panic!("expected single source, got {:?}", other),
-        },
-        other => panic!("expected single config, got {:?}", other),
-    }
-
-    // Absolute target path must remain unchanged.
-    cfg.target_curve = Some(TargetCurveConfig::Path("/abs/target.csv".into()));
-    cfg.resolve_paths(PathBuf::from("/base").as_path());
-    match &cfg.target_curve {
-        Some(TargetCurveConfig::Path(p)) => assert_eq!(p, &PathBuf::from("/abs/target.csv")),
-        other => panic!("expected path target curve, got {:?}", other),
-    }
-}
-
-#[test]
-fn policy_target_response_covers_all_presets() {
-    let reference = policy_target_response(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference.shape, TargetShape::Flat);
-    assert!(reference.broadband_precorrection);
-    assert!(reference.role_targets.is_some());
-
-    let music = policy_target_response(PerceptualPolicyPreset::Music);
-    assert_eq!(music.shape, TargetShape::Harman);
-    assert_eq!(music.slope_db_per_octave, -0.8);
-    assert_eq!(music.preference.bass_shelf_db, 1.0);
-    assert_eq!(music.preference.treble_shelf_db, -0.5);
-
-    let cinema = policy_target_response(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.shape, TargetShape::Custom);
-    let cinema_roles = cinema.role_targets.unwrap();
-    assert!(cinema_roles.cinema_x_curve_enabled);
-    assert_eq!(cinema_roles.center_dialog_boost_db, 1.5);
-
-    let night = policy_target_response(PerceptualPolicyPreset::Night);
-    assert_eq!(night.preference.bass_shelf_db, -1.5);
-    let night_roles = night.role_targets.unwrap();
-    assert_eq!(night_roles.subwoofer_bass_shelf_db, -2.0);
-
-    let speech = policy_target_response(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.slope_db_per_octave, -0.3);
-    let speech_roles = speech.role_targets.unwrap();
-    assert_eq!(speech_roles.center_dialog_boost_db, 3.0);
-    assert_eq!(speech_roles.center_dialog_low_hz, 250.0);
-    assert_eq!(speech_roles.center_dialog_high_hz, 5_000.0);
-}
-
-#[test]
-fn policy_multi_measurement_covers_all_presets() {
-    let reference = policy_multi_measurement(PerceptualPolicyPreset::Reference);
-    assert_eq!(
-        reference.strategy,
-        MultiMeasurementStrategy::MinimaxUncertainty
-    );
-
-    let music = policy_multi_measurement(PerceptualPolicyPreset::Music);
-    assert_eq!(music.strategy, MultiMeasurementStrategy::SpatialRobustness);
-
-    let cinema = policy_multi_measurement(PerceptualPolicyPreset::Cinema);
-    assert_eq!(
-        cinema.strategy,
-        MultiMeasurementStrategy::MinimaxUncertainty
-    );
-
-    let night = policy_multi_measurement(PerceptualPolicyPreset::Night);
-    assert_eq!(night.strategy, MultiMeasurementStrategy::SpatialRobustness);
-    assert_eq!(
-        night
-            .spatial_robustness
-            .as_ref()
-            .unwrap()
-            .variance_threshold_db,
-        2.5
-    );
-
-    let speech = policy_multi_measurement(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.strategy, MultiMeasurementStrategy::SpatialRobustness);
-    assert_eq!(
-        speech.bootstrap_uncertainty.as_ref().unwrap().scalarisation,
-        super::types::BootstrapScalarisation::Cvar
-    );
-    assert_eq!(
-        speech
-            .spatial_robustness
-            .as_ref()
-            .unwrap()
-            .min_correction_depth,
-        0.25
-    );
-}
-
-#[test]
-fn policy_decomposed_correction_covers_all_presets() {
-    let reference = policy_decomposed_correction(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference.early_reflection_weight, 0.1);
-    assert_eq!(reference.steady_state_weight, 0.3);
-    assert_eq!(reference.mode_correction_weight, 1.0);
-
-    let speech = policy_decomposed_correction(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.early_reflection_weight, 0.6);
-    assert_eq!(speech.steady_state_weight, 0.55);
-    assert_eq!(speech.mode_correction_weight, 1.0);
-
-    let night = policy_decomposed_correction(PerceptualPolicyPreset::Night);
-    assert_eq!(night.early_reflection_weight, 0.15);
-    assert_eq!(night.mode_correction_weight, 0.85);
-
-    let cinema = policy_decomposed_correction(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.early_reflection_weight, 0.2);
-    assert_eq!(cinema.steady_state_weight, 0.4);
-}
-
-#[test]
-fn policy_psychoacoustic_smoothing_covers_all_presets() {
-    let reference = policy_psychoacoustic_smoothing(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference, PsychoacousticSmoothingConfig::default());
-
-    let music = policy_psychoacoustic_smoothing(PerceptualPolicyPreset::Music);
-    assert_eq!(music.high_freq_n, 5);
-
-    let cinema = policy_psychoacoustic_smoothing(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.high_freq_n, 4);
-    assert_eq!(cinema.high_freq, 900.0);
-
-    let night = policy_psychoacoustic_smoothing(PerceptualPolicyPreset::Night);
-    assert_eq!(night.high_freq_n, 4);
-    assert_eq!(night.high_freq, 900.0);
-
-    let speech = policy_psychoacoustic_smoothing(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.low_freq_n, 24);
-    assert_eq!(speech.high_freq_n, 6);
-    assert_eq!(speech.low_freq, 150.0);
-    assert_eq!(speech.high_freq, 1_200.0);
-}
-
-#[test]
-fn policy_asymmetric_loss_covers_all_presets() {
-    let reference = policy_asymmetric_loss(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference, AsymmetricLossConfig::default());
-
-    let music = policy_asymmetric_loss(PerceptualPolicyPreset::Music);
-    assert_eq!(music.bass_peak_weight, 5.0);
-    assert_eq!(music.transition_freq, 200.0);
-
-    let night = policy_asymmetric_loss(PerceptualPolicyPreset::Night);
-    assert_eq!(night.bass_peak_weight, 6.0);
-    assert_eq!(night.transition_freq, 220.0);
-
-    let speech = policy_asymmetric_loss(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.bass_peak_weight, 3.0);
-    assert_eq!(speech.transition_freq, 180.0);
-}
-
-#[test]
-fn policy_smoothness_penalty_covers_all_presets() {
-    let reference = policy_smoothness_penalty(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference.tv2_weight, 0.001);
-
-    let music = policy_smoothness_penalty(PerceptualPolicyPreset::Music);
-    assert_eq!(music.tv2_weight, 0.0015);
-
-    let cinema = policy_smoothness_penalty(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.tv2_weight, 0.002);
-
-    let night = policy_smoothness_penalty(PerceptualPolicyPreset::Night);
-    assert_eq!(night.tv2_weight, 0.003);
-
-    let speech = policy_smoothness_penalty(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.tv2_weight, 0.001);
-    assert_eq!(speech.modal_weight_scale, 0.1);
-    assert_eq!(speech.exponent, 1.0);
-    assert!(speech.schroeder_hz.is_some());
-}
-
-#[test]
-fn policy_audibility_deadband_covers_all_presets() {
-    let reference = policy_audibility_deadband(PerceptualPolicyPreset::Reference);
-    let default = super::audibility_deadband_config::AudibilityDeadbandConfig::default();
-    assert_eq!(reference.bass_db, default.bass_db);
-    assert_eq!(reference.mid_db, default.mid_db);
-    assert_eq!(reference.treble_db, default.treble_db);
-
-    let music = policy_audibility_deadband(PerceptualPolicyPreset::Music);
-    assert_eq!(music.mid_db, 0.65);
-    assert_eq!(music.treble_db, 0.9);
-
-    let cinema = policy_audibility_deadband(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.mid_db, 0.75);
-    assert_eq!(cinema.treble_db, 1.1);
-
-    let night = policy_audibility_deadband(PerceptualPolicyPreset::Night);
-    assert_eq!(night.bass_db, 0.5);
-    assert_eq!(night.mid_db, 0.9);
-    assert_eq!(night.treble_db, 1.25);
-
-    let speech = policy_audibility_deadband(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.bass_db, 0.75);
-    assert_eq!(speech.mid_db, 0.5);
-    assert_eq!(speech.treble_db, 0.75);
-    assert!(!speech.disable_below_schroeder);
-}
-
-#[test]
-fn policy_high_frequency_guard_covers_all_presets() {
-    let reference = policy_high_frequency_guard(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference.extra_deadband_db, 0.75);
-    assert_eq!(reference.smoothing_n, 3);
-    assert_eq!(reference.max_q, 2.0);
-
-    let speech = policy_high_frequency_guard(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.extra_deadband_db, 0.4);
-    assert_eq!(speech.smoothing_n, 5);
-    assert_eq!(speech.max_q, 2.5);
-
-    let music = policy_high_frequency_guard(PerceptualPolicyPreset::Music);
-    assert_eq!(music.extra_deadband_db, 0.6);
-    assert_eq!(music.smoothing_n, 4);
-
-    let cinema = policy_high_frequency_guard(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.extra_deadband_db, 0.9);
-    assert_eq!(cinema.smoothing_n, 3);
-
-    let night = policy_high_frequency_guard(PerceptualPolicyPreset::Night);
-    assert_eq!(night.extra_deadband_db, 0.9);
-    assert_eq!(night.smoothing_n, 3);
-}
-
-#[test]
-fn policy_early_late_correction_covers_all_presets() {
-    let reference = policy_early_late_correction(PerceptualPolicyPreset::Reference);
-    assert_eq!(reference.early_cue_risk_db, -22.0);
-
-    let music = policy_early_late_correction(PerceptualPolicyPreset::Music);
-    assert_eq!(music.early_cue_risk_db, -18.0);
-
-    let cinema = policy_early_late_correction(PerceptualPolicyPreset::Cinema);
-    assert_eq!(cinema.early_cue_risk_db, -18.0);
-
-    let night = policy_early_late_correction(PerceptualPolicyPreset::Night);
-    assert_eq!(night.early_cue_risk_db, -20.0);
-
-    let speech = policy_early_late_correction(PerceptualPolicyPreset::Speech);
-    assert_eq!(speech.early_cue_risk_db, -14.0);
-}
-
-#[test]
-fn all_default_functions_return_sane_values() {
-    // Exercise every simple default_* function in default.rs to cover all
-    // the missed one-liner bodies.
-    let _ = super::default::default_config_version();
-    let _ = super::default::default_bass_management_enabled();
-    let _ = super::default::default_redirect_bass();
-    let _ = super::default::default_lfe_channel();
-    let _ = super::default::default_lfe_playback_gain_db();
-    let _ = super::default::default_sub_headroom_margin_db();
-    let _ = super::default::default_max_sub_boost_db();
-    let _ = super::default::default_optimize_bass_groups();
-    let _ = super::default::default_bass_headroom_model_kind();
-    let _ = super::default::default_lr_correlation();
-    let _ = super::default::default_lcr_correlation();
-    let _ = super::default::default_surround_height_correlation();
-    let _ = super::default::default_tilt_slope();
-    let _ = super::default::default_tilt_reference_freq();
-    let _ = super::default::default_bass_shelf_freq();
-    let _ = super::default::default_treble_shelf_freq();
-    let _ = super::default::default_role_targets_enabled();
-    let _ = super::default::default_center_dialog_low_hz();
-    let _ = super::default::default_center_dialog_high_hz();
-    let _ = super::default::default_cinema_reference_distance_m();
-    let _ = super::default::default_cinema_x_curve_start_hz();
-    let _ = super::default::default_pre_ringing_threshold();
-    let _ = super::default::default_pre_ringing_time();
-    let _ = super::default::default_fir_taps();
-    let _ = super::default::default_fir_phase();
-    let _ = super::default::default_phase_smoothing();
-    let _ = super::default::default_mixed_phase_fir_length();
-    let _ = super::default::default_mixed_phase_spatial_depth();
-    let _ = super::default::default_mask_smoothing();
-    let _ = super::default::default_crossover_freq();
-    let _ = super::default::default_crossover_type();
-    let _ = super::default::default_fir_band();
-    let _ = super::default::default_true();
-    let _ = super::default::default_f3_reference_min_hz();
-    let _ = super::default::default_f3_reference_max_hz();
-    let _ = super::default::default_filter_order();
-    let _ = super::default::default_margin_octaves();
-    let _ = super::default::default_low_freq_max_q();
-    let _ = super::default::default_high_freq_max_q();
-    let _ = super::default::default_schroeder_freq();
-    let _ = super::default::default_phase_min_freq();
-    let _ = super::default::default_phase_max_freq();
-    let _ = super::default::default_max_delay_ms();
-    let _ = super::default::default_max_deviation_db();
-    let _ = super::default::default_all_channel_multiseat_enabled();
-    let _ = super::default::default_all_channel_multiseat_strategy();
-    let _ = super::default::default_primary_seat_weight();
-    let _ = super::default::default_multiseat_per_sub_peq();
-    let _ = super::default::default_multiseat_global_eq();
-    let _ = super::default::default_channel_matching_threshold();
-    let _ = super::default::default_channel_matching_max_filters();
-    let _ = super::default::default_gd_max_delay_ms();
-    let _ = super::default::default_gd_ap_per_channel();
-    let _ = super::default::default_gd_ap_min_q();
-    let _ = super::default::default_gd_ap_max_q();
-    let _ = super::default::default_gd_coherence_threshold();
-    let _ = super::default::default_gd_min_improvement_db();
-    let _ = super::default::default_gd_max_iter();
-    let _ = super::default::default_gd_popsize();
-    let _ = super::default::default_gd_tol();
-    let _ = super::default::default_sub_num_filters();
-    let _ = super::default::default_sub_max_db();
-    let _ = super::default::default_sub_min_db();
-    let _ = super::default::default_sub_max_q();
-    let _ = super::default::default_modal_weight_scale();
-    let _ = super::default::default_smoothness_exponent();
-    let _ = super::default::default_variance_threshold();
-    let _ = super::default::default_transition_width();
-    let _ = super::default::default_min_correction_depth();
-    let _ = super::default::default_mask_smoothing_octaves();
-    let _ = super::default::default_bootstrap_num_resamples();
-    let _ = super::default::default_bootstrap_alpha();
-    let _ = super::default::default_bootstrap_seed();
-    let _ = super::default::default_bootstrap_cvar_alpha();
-    let _ = super::default::default_gaussian_truncation_sigmas();
-    let _ = super::default::default_area_inner_maxiter();
-    let _ = super::default::default_area_cvar_alpha();
-    let _ = super::default::default_idw_power();
-    let _ = super::default::default_variance_lambda();
-    let _ = super::default::default_decomposed_schroeder();
-    let _ = super::default::default_decomposed_min_q();
-    let _ = super::default::default_decomposed_prominence();
-    let _ = super::default::default_decomposed_mode_weight();
-    let _ = super::default::default_decomposed_reflection_weight();
-    let _ = super::default::default_decomposed_steady_weight();
-    let _ = super::default::default_fdw_cycles();
-    let _ = super::default::default_fdw_min_window_ms();
-    let _ = super::default::default_fdw_max_window_ms();
-    let _ = super::default::default_fdw_smoothing_octaves();
-    let _ = super::default::default_cea2034_version();
-    let _ = super::default::default_nearfield_threshold();
-    let _ = super::default::default_cea2034_num_filters();
-    let _ = super::default::default_cea2034_max_q();
-    let _ = super::default::default_cea2034_max_db();
-    let _ = super::default::default_cea2034_min_db();
-    let _ = super::default::default_auto_min_filters();
-    let _ = super::default::default_auto_max_filters();
-    let _ = super::default::default_deadband_bass_db();
-    let _ = super::default::default_deadband_mid_db();
-    let _ = super::default::default_deadband_treble_db();
-    let _ = super::default::default_deadband_bass_mid_hz();
-    let _ = super::default::default_deadband_mid_treble_hz();
-    let _ = super::default::default_high_freq_guard_start_hz();
-    let _ = super::default::default_high_freq_extra_deadband_db();
-    let _ = super::default::default_high_freq_smoothing_n();
-    let _ = super::default::default_high_freq_guard_max_q();
-    let _ = super::default::default_direct_window_ms();
-    let _ = super::default::default_early_window_ms();
-    let _ = super::default::default_late_window_ms();
-    let _ = super::default::default_early_cue_risk_db();
-    let _ = super::default::default_validation_lufs();
-    let _ = super::default::default_loss_type();
-    let _ = super::default::default_algorithm();
-    let _ = super::default::default_strategy();
-    let _ = super::default::default_peq_model();
-    let _ = super::default::default_num_filters();
-    let _ = super::default::default_min_filter_improvement();
-    let _ = super::default::default_elimination_threshold();
-    let _ = super::default::default_min_q();
-    let _ = super::default::default_max_q();
-    let _ = super::default::default_min_db();
-    let _ = super::default::default_max_db();
-    let _ = super::default::default_min_freq();
-    let _ = super::default::default_max_freq();
-    let _ = super::default::default_max_iter();
-    let _ = super::default::default_population();
-    let _ = super::default::default_refine();
-    let _ = super::default::default_local_algo();
-    let _ = super::default::default_psychoacoustic();
-    let _ = super::default::default_smooth_n();
-    let _ = super::default::default_asymmetric_loss();
-    let _ = super::default::default_tolerance();
-    let _ = super::default::default_atolerance();
-    let _ = super::default::default_ctc_matrix_source();
-    let _ = super::default::default_ctc_window_type();
-    let _ = super::default::default_ctc_window_start_ms();
-    let _ = super::default::default_ctc_window_length_ms();
-    let _ = super::default::default_ctc_window_fade_ms();
-    let _ = super::default::default_ctc_beta_db();
-    let _ = super::default::default_ctc_max_gain_db();
-    let _ = super::default::default_ctc_fir_taps();
-    let _ = super::default::default_ctc_robustness();
-    let _ = super::default::default_ctc_include_room_eq_dsp();
-    let _ = super::default::default_ctc_hrtf_distance_m();
-    let _ = super::default::default_ctc_fdw_cycles();
-    let _ = super::default::default_ctc_fdw_min_ms();
-    let _ = super::default::default_ctc_fdw_max_ms();
-    let _ = super::default::default_ctc_max_harmonic();
-    let _ = super::default::default_ctc_harmonic_window_ms();
-    let _ = super::default::default_ctc_minimax_iterations();
-}
+include!("tests/core.rs");
+include!("tests/roundtrip.rs");
+include!("tests/policies.rs");
