+++ src/roomeq/types/config/tests/core.rs
@@ -1,0 +1,387 @@
+#[test]
+fn spl_calibration_roundtrip_and_helpers() {
+    let cal = SplCalibration {
+        reported_db_spl: 85.0,
+        reference_freq_hz: 1000.0,
+        peak_sample_level: 0.5,
+        spl_offset_db: 85.0 - 20.0 * 0.5_f32.log10(),
+    };
+    // Round-trip through JSON.
+    let json = serde_json::to_string(&cal).unwrap();
+    let back: SplCalibration = serde_json::from_str(&json).unwrap();
+    assert_eq!(back, cal);
+    // `dbspl_for_peak_level` at the calibration peak must return the
+    // reported dBSPL (within f32 rounding).
+    let recovered = cal.dbspl_for_peak_level(cal.peak_sample_level);
+    assert!((recovered - cal.reported_db_spl).abs() < 1e-3);
+    // `peak_level_for_dbspl` is the inverse.
+    let peak = cal.peak_level_for_dbspl(cal.reported_db_spl);
+    assert!((peak - cal.peak_sample_level).abs() < 1e-5);
+    // Clamp: asking for a dBSPL below the representable range
+    // should return 0.0 rather than a negative peak level.
+    assert_eq!(cal.peak_level_for_dbspl(-1000.0), 0.0);
+}
+
+#[test]
+fn recording_configuration_accepts_gd_v2_fields() {
+    let cfg_json = serde_json::json!({
+        "bass_octave_duration_s": 3.0,
+        "pre_silence_s": 2.0,
+        "post_silence_s": 4.0,
+        "sweep_level_db_spl": 85.0,
+        "num_sweeps": 4,
+        "coherence_threshold": 0.9,
+        "bass_probe_freq_hz": 30.0,
+        "bass_probe_duration_s": 2.0,
+        "mic_phase_calibration_path": "/tmp/mic_phase.csv",
+        "spl_calibration": {
+            "reported_db_spl": 85.0,
+            "reference_freq_hz": 1000.0,
+            "peak_sample_level": 0.5,
+            "spl_offset_db": 91.02
+        },
+        "recording_seed": 42
+    });
+    let cfg: RecordingConfiguration = serde_json::from_value(cfg_json).unwrap();
+    assert_eq!(cfg.bass_octave_duration_s, Some(3.0));
+    assert_eq!(cfg.num_sweeps, Some(4));
+    assert_eq!(cfg.coherence_threshold, Some(0.9));
+    assert_eq!(cfg.bass_probe_freq_hz, Some(30.0));
+    assert_eq!(cfg.bass_probe_duration_s, Some(2.0));
+    assert_eq!(
+        cfg.mic_phase_calibration_path.as_deref(),
+        Some("/tmp/mic_phase.csv")
+    );
+    let cal = cfg.spl_calibration.expect("spl_calibration populated");
+    assert!((cal.reported_db_spl - 85.0).abs() < 1e-6);
+    assert_eq!(cfg.recording_seed, Some(42));
+}
+
+#[test]
+fn recording_configuration_legacy_json_still_loads() {
+    // A session written before GD-1a only knows the pre-existing
+    // fields. GD-Opt v2 metadata must default to `None` rather
+    // than failing deserialization — the confidence gate is what
+    // converts `None` into `Advisory::GdOptDegradedPhase` at
+    // optimization time.
+    let legacy_json = serde_json::json!({
+        "signal_type": "Sweep",
+        "signal_duration_secs": 10.0,
+        "sweep_start_freq": 20.0,
+        "sweep_end_freq": 20000.0
+    });
+    let cfg: RecordingConfiguration = serde_json::from_value(legacy_json).unwrap();
+    assert!(cfg.bass_octave_duration_s.is_none());
+    assert!(cfg.num_sweeps.is_none());
+    assert!(cfg.coherence_threshold.is_none());
+    assert!(cfg.bass_probe_freq_hz.is_none());
+    assert!(cfg.bass_probe_duration_s.is_none());
+    assert!(cfg.mic_phase_calibration_path.is_none());
+    assert!(cfg.mic_phase_calibration_paths.is_none());
+    assert!(cfg.spl_calibration.is_none());
+    assert!(cfg.recording_seed.is_none());
+}
+
+#[test]
+fn bass_anchor_results_legacy_migrates_v1_bass_cycles_to_duration() {
+    // Pre-v2 schema: bass_cycles + bass_freq_hz, no bass_duration_s.
+    let legacy = serde_json::json!({
+        "channels": [],
+        "sample_rate": 48_000_u32,
+        "bass_freq_hz": 30.0_f32,
+        "bass_cycles": 6_u16,
+    });
+    let r: BassAnchorResultsLegacy = serde_json::from_value(legacy).unwrap();
+    // 6 cycles at 30 Hz = 0.2 s.
+    assert!(
+        (r.bass_duration_s - 0.2).abs() < 1e-6,
+        "expected 0.2 s migrated from 6 cycles @ 30 Hz, got {}",
+        r.bass_duration_s
+    );
+    assert_eq!(r.sample_rate, 48_000);
+    assert_eq!(r.bass_freq_hz, 30.0);
+}
+
+#[test]
+fn bass_anchor_results_prefers_explicit_duration_when_both_present() {
+    let mixed = serde_json::json!({
+        "channels": [],
+        "sample_rate": 48_000_u32,
+        "bass_freq_hz": 30.0_f32,
+        "bass_cycles": 6_u16,
+        "bass_duration_s": 2.5_f32,
+    });
+    let r: BassAnchorResultsLegacy = serde_json::from_value(mixed).unwrap();
+    assert!((r.bass_duration_s - 2.5).abs() < 1e-6);
+}
+
+#[test]
+fn bass_anchor_results_v2_round_trips() {
+    let v2 = serde_json::json!({
+        "channels": [],
+        "sample_rate": 48_000_u32,
+        "bass_freq_hz": 30.0_f32,
+        "bass_duration_s": 2.0_f32,
+    });
+    let r: BassAnchorResultsLegacy = serde_json::from_value(v2).unwrap();
+    assert_eq!(r.bass_duration_s, 2.0);
+}
+
+#[test]
+fn target_shape_canonical_wire_format() {
+    // Pins the on-the-wire string for every TargetShape variant.
+    // `from_measurement` (underscore) is the sole canonical form,
+    // matching bin/roomeq/input_schema.json and INPUT_FORMAT.md.
+    let cases = [
+        (TargetShape::Flat, "\"flat\""),
+        (TargetShape::Harman, "\"harman\""),
+        (TargetShape::Custom, "\"custom\""),
+        (TargetShape::File, "\"file\""),
+        (TargetShape::FromMeasurement, "\"from_measurement\""),
+    ];
+    for (variant, expected) in cases {
+        let serialized = serde_json::to_string(&variant).unwrap();
+        assert_eq!(serialized, expected, "serialize {variant:?}");
+        let round_tripped: TargetShape = serde_json::from_str(&serialized).unwrap();
+        assert_eq!(round_tripped, variant, "round-trip {variant:?}");
+    }
+    // Old canonical form used before the snake_case switch must no
+    // longer deserialize — a paranoid guard against accidental
+    // reintroduction of the `#[serde(alias = "from_measurement")]`
+    // back-compat shim.
+    assert!(serde_json::from_str::<TargetShape>("\"frommeasurement\"").is_err());
+}
+
+#[test]
+fn test_optimizer_config_default_has_decomposed_correction_enabled() {
+    let config = OptimizerConfig::default();
+    let dc = config
+        .decomposed_correction
+        .expect("decomposed_correction should be Some by default");
+    assert!(
+        dc.enabled,
+        "decomposed_correction should be enabled by default"
+    );
+    assert_eq!(dc.schroeder_freq, 250.0);
+    assert_eq!(dc.steady_state_weight, 0.4);
+}
+
+#[test]
+fn test_optimizer_config_default_algorithm_is_cmaes() {
+    let config = OptimizerConfig::default();
+    assert_eq!(config.algorithm, "autoeq:cmaes");
+}
+
+#[test]
+fn test_optimizer_config_resolves_new_acoustic_defaults() {
+    let config = OptimizerConfig::default();
+    assert_eq!(
+        config.psychoacoustic_smoothing_config(),
+        PsychoacousticSmoothingConfig::default()
+    );
+    assert_eq!(
+        config.asymmetric_loss_config(),
+        AsymmetricLossConfig::default()
+    );
+
+    let custom_smoothing = PsychoacousticSmoothingConfig {
+        low_freq_n: 24,
+        high_freq_n: 3,
+        low_freq: 80.0,
+        high_freq: 800.0,
+    };
+    let custom_asym = AsymmetricLossConfig {
+        peak_weight: 3.0,
+        dip_weight: 0.5,
+        bass_peak_weight: 6.0,
+        bass_dip_weight: 0.25,
+        transition_freq: 180.0,
+    };
+    let config = OptimizerConfig {
+        psychoacoustic_smoothing: Some(custom_smoothing),
+        asymmetric_loss_config: Some(custom_asym),
+        ..Default::default()
+    };
+    assert_eq!(config.psychoacoustic_smoothing_config(), custom_smoothing);
+    assert_eq!(config.asymmetric_loss_config(), custom_asym);
+}
+
+#[test]
+fn perceptual_policy_preserves_legacy_defaults_when_absent() {
+    let mut config = OptimizerConfig::default();
+    config.apply_perceptual_policy_defaults();
+
+    assert!(config.perceptual_policy.is_none());
+    assert!(config.audibility_deadband.is_none());
+    assert!(config.high_frequency_correction.is_none());
+    assert!(config.early_late_correction.is_none());
+    assert!(config.validation_bundle.is_none());
+    assert_eq!(config.loss_type, "flat");
+    assert_eq!(config.max_freq, default_max_freq());
+}
+
+#[test]
+fn music_policy_fills_perceptual_defaults() {
+    let mut config = OptimizerConfig {
+        perceptual_policy: Some(PerceptualPolicyConfig {
+            preset: PerceptualPolicyPreset::Music,
+            ..Default::default()
+        }),
+        ..Default::default()
+    };
+
+    config.apply_perceptual_policy_defaults();
+
+    assert_eq!(config.loss_type, "epa");
+    assert_eq!(
+        config.target_response.as_ref().unwrap().shape,
+        TargetShape::Harman
+    );
+    assert!(config.audibility_deadband_config().is_some());
+    assert!(config.high_frequency_correction.is_some());
+    assert!(config.early_late_correction_config().is_some());
+    assert!(config.validation_bundle_config().is_some());
+    assert_eq!(
+        config.multi_measurement.as_ref().unwrap().strategy,
+        MultiMeasurementStrategy::SpatialRobustness
+    );
+}
+
+#[test]
+fn high_frequency_guard_caps_default_q_but_preserves_explicit_q() {
+    let mut default_q = OptimizerConfig {
+        max_freq: 8_000.0,
+        high_frequency_correction: Some(HighFrequencyCorrectionConfig::default()),
+        ..Default::default()
+    };
+    default_q.apply_high_frequency_correction_defaults(false);
+    assert_eq!(default_q.max_q, default_high_freq_guard_max_q());
+    assert_eq!(
+        default_q.psychoacoustic_smoothing.unwrap().high_freq_n,
+        default_high_freq_smoothing_n()
+    );
+
+    let mut explicit_q = OptimizerConfig {
+        max_freq: 8_000.0,
+        max_q: 4.5,
+        high_frequency_correction: Some(HighFrequencyCorrectionConfig::default()),
+        ..Default::default()
+    };
+    explicit_q.apply_high_frequency_correction_defaults(false);
+    assert_eq!(explicit_q.max_q, 4.5);
+
+    explicit_q.apply_high_frequency_correction_defaults(true);
+    assert_eq!(explicit_q.max_q, default_high_freq_guard_max_q());
+}
+
+#[test]
+fn perceptual_policy_tightens_generated_high_frequency_smoothing() {
+    let mut config = OptimizerConfig {
+        max_freq: 8_000.0,
+        perceptual_policy: Some(PerceptualPolicyConfig {
+            preset: PerceptualPolicyPreset::Music,
+            ..Default::default()
+        }),
+        ..Default::default()
+    };
+
+    config.apply_perceptual_policy_defaults();
+
+    let smoothing = config.psychoacoustic_smoothing.unwrap();
+    assert_eq!(
+        smoothing.high_freq_n,
+        policy_high_frequency_guard(PerceptualPolicyPreset::Music).smoothing_n
+    );
+    assert!(smoothing.high_freq <= default_high_freq_guard_start_hz());
+}
+
+#[test]
+fn test_decomposed_correction_serde_config_default() {
+    let dc = DecomposedCorrectionSerdeConfig::default();
+    assert!(dc.enabled);
+    assert_eq!(dc.schroeder_freq, 250.0);
+    assert_eq!(dc.steady_state_weight, 0.4);
+    assert_eq!(dc.min_mode_q, 3.0);
+    assert_eq!(dc.min_mode_prominence_db, 3.0);
+    assert_eq!(dc.mode_correction_weight, 1.0);
+    assert_eq!(dc.early_reflection_weight, 0.3);
+    assert!(dc.fdw_enabled);
+    assert_eq!(dc.fdw_cycles, 8.0);
+    assert_eq!(dc.fdw_min_window_ms, 3.0);
+    assert_eq!(dc.fdw_max_window_ms, 500.0);
+    assert_eq!(dc.fdw_smoothing_octaves, 1.0 / 24.0);
+}
+
+#[test]
+fn test_channel_matching_config_defaults() {
+    let cfg = ChannelMatchingConfig::default();
+    assert!(cfg.enabled);
+    assert_eq!(cfg.threshold_db, 0.75);
+    assert_eq!(cfg.max_filters, 5);
+}
+
+#[test]
+fn test_max_boost_envelope_interpolation() {
+    let mut config = OptimizerConfig::default();
+
+    // Without envelope, falls back to max_db
+    assert_eq!(config.max_boost_at_freq(100.0), config.max_db);
+
+    // Set an envelope: generous bass boost tapering to zero
+    config.max_boost_envelope = Some(vec![
+        (20.0, 6.0),
+        (200.0, 4.0),
+        (1000.0, 2.0),
+        (8000.0, 0.0),
+    ]);
+
+    // At exact envelope points
+    assert!((config.max_boost_at_freq(20.0) - 6.0).abs() < 1e-10);
+    assert!((config.max_boost_at_freq(200.0) - 4.0).abs() < 1e-10);
+    assert!((config.max_boost_at_freq(1000.0) - 2.0).abs() < 1e-10);
+    assert!((config.max_boost_at_freq(8000.0) - 0.0).abs() < 1e-10);
+
+    // Below first point: clamp to first value
+    assert!((config.max_boost_at_freq(10.0) - 6.0).abs() < 1e-10);
+
+    // Above last point: clamp to last value
+    assert!((config.max_boost_at_freq(16000.0) - 0.0).abs() < 1e-10);
+
+    // Between 200Hz and 1000Hz: log-frequency interpolation
+    // Geometric midpoint of 200 and 1000 is sqrt(200*1000) ~ 447Hz
+    let mid_freq = (200.0_f64 * 1000.0).sqrt();
+    let mid_boost = config.max_boost_at_freq(mid_freq);
+    // At geometric midpoint, t = 0.5, so interpolated value = 4.0 + 0.5*(2.0-4.0) = 3.0
+    assert!(
+        (mid_boost - 3.0).abs() < 1e-6,
+        "geometric midpoint should give 3.0 dB, got {:.6}",
+        mid_boost
+    );
+}
+
+#[test]
+fn config_types_default_roundtrip() {
+    use super::*;
+
+    let _ = AreaPriorKind::default();
+    let _ = AreaQuadratureKind::default();
+    let _ = AreaScalarisationKind::default();
+    let _ = BootstrapScalarisation::default();
+    let _ = CtcConfig::default();
+    let _ = CtcRegularizationConfig::default();
+    let _ = CtcWindowConfig::default();
+    let _ = MixedModeConfig::default();
+    let _ = SubOptimizerConfig::default();
+
+    let _ = serde_json::to_string(&CtcConfig::default()).unwrap();
+    let _ = serde_json::to_string(&MixedModeConfig::default()).unwrap();
+    let _ = serde_json::to_string(&SubOptimizerConfig::default()).unwrap();
+
+    let quad = AreaQuadratureKind::Sobol {
+        num_points: 16,
+        seed: 42,
+    };
+    let json = serde_json::to_string(&quad).unwrap();
+    let back: AreaQuadratureKind = serde_json::from_str(&json).unwrap();
+    assert_eq!(back, quad);
+}
