+++ tests/roomeq_generated_data_test.rs
@@ -2,6 +2,10 @@
 //!
 //! These tests load pre-computed FEM simulation data and verify that
 //! the roomeq optimizer can improve the simulated frequency responses.
+//!
+//! The smallest stereo convergence case runs in the default suite. Run the
+//! complete long-running convergence and multimode matrix with:
+//! `cargo test -p autoeq --test roomeq_generated_data_test -- --ignored --test-threads=1`
 
 #[path = "roomeq_generated_data_test/consts.rs"]
 mod consts;
