--- tests/validator_bounds_test.rs
@@ -1,101 +1,0 @@
-//! Unit test for the B3 bounds-check helper.
-//!
-//! `warn_if_optimizer_bounds_exceed_data` is a private-in-crate helper that
-//! emits a `log::warn!` when the optimizer's `min_freq`/`max_freq` fall
-//! outside the measurement's frequency range. Rather than plumb a log
-//! capture harness, we test the observable side effect — whether the helper
-//! is called with the right preconditions — by driving the full optimizer
-//! with an in-memory curve and asserting the downstream behaviour stays
-//! consistent. Since the helper itself is private, we validate the *guards*
-//! that drive it (data-min vs opt.min_freq) through a focused unit via the
-//! `log` crate's `Log` trait.
-
-fn curve(freqs: Vec<f64>) -> autoeq::Curve {
-    let n = freqs.len();
-    autoeq::Curve {
-        freq: ndarray::Array1::from(freqs),
-        spl: ndarray::Array1::zeros(n),
-        phase: None,
-        ..Default::default()
-    }
-}
-
-// The helper is `pub(super)` within the crate, so we exercise it through the
-// public re-export path we've added for tests: it lives in
-// `autoeq::roomeq::optimize`, which is a private module. To reach it from an
-// integration test, we re-implement the exact observable contract here by
-// driving a thin shim: we call the real public API that triggers it.
-// Since adding a shim would be invasive, we validate the contract via a
-// direct call path that is already public — the `log::warn!` output when
-// `process_single_speaker` runs with out-of-range bounds would only fire
-// during optimisation, which is too heavy for a unit test. Instead we pin
-// the equivalent logic inline so that if the helper's semantics ever drift
-// (e.g. tolerance changes from 5 % to 10 % of a log-decade), this test
-// catches the divergence.
-
-#[test]
-fn bounds_helper_triggers_when_min_freq_below_data() {
-    // Exact replica of the helper's decision thresholds, so this test acts
-    // as a contract check against the helper.
-    fn reference_check(curve: &autoeq::Curve, min_freq: f64, max_freq: f64) -> (bool, bool) {
-        let data_min = curve.freq[0];
-        let data_max = curve.freq[curve.freq.len() - 1];
-        let log_margin = 0.05;
-        let min_tol = data_min * 10_f64.powf(-log_margin);
-        let max_tol = data_max * 10_f64.powf(log_margin);
-        (min_freq < min_tol, max_freq > max_tol)
-    }
-
-    // Data covers 100 Hz .. 10 kHz.
-    let c = curve(vec![100.0, 1000.0, 10_000.0]);
-
-    // min_freq = 50 Hz is well below 100 Hz → should trigger.
-    let (below, above) = reference_check(&c, 50.0, 5000.0);
-    assert!(below, "min_freq=50 should trigger below-data warning");
-    assert!(
-        !above,
-        "max_freq=5000 should not trigger above-data warning"
-    );
-
-    // max_freq = 20 kHz is well above 10 kHz → should trigger.
-    let (below, above) = reference_check(&c, 200.0, 20_000.0);
-    assert!(!below, "min_freq=200 should not trigger below-data warning");
-    assert!(above, "max_freq=20000 should trigger above-data warning");
-
-    // Bounds inside the data range and within tolerance → neither.
-    let (below, above) = reference_check(&c, 100.0, 10_000.0);
-    assert!(!below);
-    assert!(!above);
-
-    // Slight rounding — opt.min_freq=99, data_min=100 is within 5% log tolerance (~12 %).
-    // 100 * 10^-0.05 ≈ 89.1; 99 > 89.1 so no warning. Pins the tolerance.
-    let (below, above) = reference_check(&c, 99.0, 10_000.0);
-    assert!(
-        !below,
-        "99 Hz vs 100 Hz data_min must be within 5 % log-axis tolerance"
-    );
-    assert!(!above);
-}
-
-#[test]
-fn bounds_helper_handles_empty_curve() {
-    // Reference check should degrade gracefully when the curve is empty.
-    // We replicate by calling the invariant directly: an empty freq array
-    // must not panic and must not trigger warnings.
-    fn reference_check(curve: &autoeq::Curve, min_freq: f64, max_freq: f64) -> (bool, bool) {
-        if curve.freq.is_empty() {
-            return (false, false);
-        }
-        let data_min = curve.freq[0];
-        let data_max = curve.freq[curve.freq.len() - 1];
-        let log_margin = 0.05;
-        let min_tol = data_min * 10_f64.powf(-log_margin);
-        let max_tol = data_max * 10_f64.powf(log_margin);
-        (min_freq < min_tol, max_freq > max_tol)
-    }
-
-    let empty = curve(vec![]);
-    let (below, above) = reference_check(&empty, 10.0, 20000.0);
-    assert!(!below);
-    assert!(!above);
-}
